refactor: extract engine.ts module-level helpers into focused modules
Phase 1 of the engine.ts decomposition (#867). Pure code movement — no
behavior changes. The ~3,100-line loose-helper layer that preceded the
LcmContextEngine class moves into twelve cohesive modules:

- value-utils: generic value/format helpers
- assemble-debug: prefix-change + overflow diagnostics formatting
- replay-metadata: raw replay id/signature extraction + canonicalization
- message-content: structured-text extraction, part building, storage
  normalization, injected-delivery detection
- prompt-recall: recall cue snippet helpers
- message-signatures: identity/signature builders
- token-accounting: content token estimates + runtime usage extraction
- tool-pairing: tool call/result index pairing (consumed by live-coverage)
- live-coverage: volatile live-input coverage matching
- assemble-fallback: budget clamps + degraded/fork-bounded fallbacks
- bootstrap-budget: bootstrap import trimming
- heartbeat-filter: synthetic heartbeat detection/filtering

engine.ts shrinks 12,886 -> ~9,600 lines and now imports these helpers;
its no-longer-used imports are pruned. Full suite: 1279 tests passing.
