+++ src/bootstrap-budget.ts
@@ -1,0 +1,72 @@
+/**
+ * Bootstrap import budgeting: cap and trim transcript messages imported at session start.
+ *
+ * Extracted from engine.ts (Phase 1 of the engine decomposition).
+ */
+import type { LcmConfig } from "./db/config.js";
+import { toStoredMessage } from "./message-content.js";
+import type { AgentMessage } from "./openclaw-bridge.js";
+
+/**
+ * Resolve the first-time bootstrap token budget.
+ *
+ * When unset, bootstrap keeps a modest suffix of the parent session rather than
+ * inheriting the full raw history into a brand-new conversation.
+ */
+export function resolveBootstrapMaxTokens(config: Pick<LcmConfig, "bootstrapMaxTokens" | "leafChunkTokens">): number {
+  if (
+    typeof config.bootstrapMaxTokens === "number" &&
+    Number.isFinite(config.bootstrapMaxTokens) &&
+    config.bootstrapMaxTokens > 0
+  ) {
+    return Math.floor(config.bootstrapMaxTokens);
+  }
+
+  const leafChunkTokens =
+    typeof config.leafChunkTokens === "number" &&
+    Number.isFinite(config.leafChunkTokens) &&
+    config.leafChunkTokens > 0
+      ? Math.floor(config.leafChunkTokens)
+      : 40_000;
+  return Math.max(6000, Math.floor(leafChunkTokens * 0.3));
+}
+
+/**
+ * Keep only the newest bootstrap messages that fit within the token budget.
+ *
+ * The newest message is always preserved so a fork never starts empty when the
+ * parent transcript has any recoverable content at all.
+ */
+export function trimBootstrapMessagesToBudget(messages: AgentMessage[], maxTokens: number): AgentMessage[] {
+  if (messages.length === 0) {
+    return [];
+  }
+
+  const safeMaxTokens = Number.isFinite(maxTokens) ? Math.floor(maxTokens) : 0;
+  if (safeMaxTokens <= 0) {
+    return [messages[messages.length - 1]!];
+  }
+
+  const kept: AgentMessage[] = [];
+  let totalTokens = 0;
+
+  for (let index = messages.length - 1; index >= 0; index -= 1) {
+    const message = messages[index]!;
+    const tokenCount = toStoredMessage(message).tokenCount;
+    if (kept.length > 0 && totalTokens + tokenCount > safeMaxTokens) {
+      break;
+    }
+    kept.push(message);
+    totalTokens += tokenCount;
+  }
+
+  // If a single oversized tail message exceeds the budget, return empty
+  // rather than silently bypassing the budget cap. An empty bootstrap is
+  // safer than an exploding one.
+  if (kept.length === 1 && totalTokens > safeMaxTokens) {
+    return [];
+  }
+
+  kept.reverse();
+  return kept;
+}
