+++ src/engine.ts
@@ -1,10 +1,8 @@
 import { createHash, randomUUID } from "node:crypto";
-import { createReadStream, readFileSync, statSync } from "node:fs";
+import { readFileSync } from "node:fs";
 import { mkdir, open, stat, writeFile } from "node:fs/promises";
-import type { FileHandle } from "node:fs/promises";
-import { join, resolve as resolvePath } from "node:path";
+import { join } from "node:path";
 import type { DatabaseSync } from "node:sqlite";
-import { createInterface } from "node:readline";
 import type {
   ContextEngine,
   ContextEngineInfo,
@@ -17,15 +15,7 @@ import type {
   SubagentEndReason,
   SubagentSpawnPreparation,
 } from "./openclaw-bridge.js";
-import {
-  blockFromPart,
-  contentFromParts,
-  ContextAssembler,
-  pickToolCallId,
-  pickToolIsError,
-  pickToolName,
-  type AssemblyOverflowDiagnostics,
-} from "./assembler.js";
+import { contentFromParts, ContextAssembler, pickToolCallId, pickToolIsError, pickToolName } from "./assembler.js";
 import { CompactionEngine, type CompactionConfig } from "./compaction.js";
 import type { LcmConfig } from "./db/config.js";
 import { getLcmDbFeatures } from "./db/features.js";
@@ -55,17 +45,8 @@ import {
   type ConversationCompactionTelemetryRecord,
   type CacheState,
 } from "./store/compaction-telemetry-store.js";
-import {
-  CompactionMaintenanceStore,
-  type ConversationCompactionMaintenanceRecord,
-} from "./store/compaction-maintenance-store.js";
-import {
-  ConversationStore,
-  type ConversationRecord,
-  type CreateMessagePartInput,
-  type MessagePartRecord,
-  type MessagePartType,
-} from "./store/conversation-store.js";
+import { CompactionMaintenanceStore } from "./store/compaction-maintenance-store.js";
+import { ConversationStore, type ConversationRecord } from "./store/conversation-store.js";
 import { buildMessageIdentityHash } from "./store/message-identity.js";
 import { FocusBriefStore, type FocusBriefRecord } from "./store/focus-brief-store.js";
 import { SummaryStore, type ContextItemRecord } from "./store/summary-store.js";
@@ -78,37 +59,28 @@ import {
   type LcmSummarizeFn,
 } from "./summarize.js";
 import type { CompleteFn, LcmDependencies, StartupSessionFileCandidate } from "./types.js";
-import {
-  estimateSerializedMessageTokens,
-  estimateSerializedMessagesTokens,
-  estimateTokens,
-} from "./estimate-tokens.js";
+import { estimateTokens } from "./estimate-tokens.js";
 import { buildDeterministicFallbackSummary } from "./summary-fallback.js";
 import { createLcmDatabaseBackup } from "./plugin/lcm-db-backup.js";
 import {
   DatabaseTransactionTimeoutError,
   withExclusiveDatabaseLock,
 } from "./transaction-mutex.js";
-import { sanitizeToolUseResultPairing } from "./transcript-repair.js";
-import {
-  extractBootstrapMessageCandidate,
-  getTranscriptEntryId,
-  getTranscriptEntryMeta,
-  isBootstrapMessage,
-  parseBootstrapJsonl,
-  readAppendedLeafPathMessages,
-  readLastJsonlEntryBeforeOffset,
-  readLeafPathMessages,
-  readLeafPathRawEntries,
-  readSessionParentSessionReference,
-  readTranscriptHeader,
-  type TranscriptRawEntry,
-} from "./transcript.js";
+import { getTranscriptEntryId, getTranscriptEntryMeta, readAppendedLeafPathMessages, readLastJsonlEntryBeforeOffset, readLeafPathMessages, readLeafPathRawEntries, readSessionParentSessionReference, readTranscriptHeader, type TranscriptRawEntry } from "./transcript.js";
 import { resolveEpochRoute, selectEntryIdTail, transcriptImportCap } from "./reconcile-plan.js";
+import { describeAssembledPrefixChange, formatOverflowDiagnosticsForLog, shouldLogOverflowDiagnostics, type AssemblePrefixSnapshot, type BootstrapImportObservation } from "./assemble-debug.js";
+import { appendForkBoundedLiveSuffixWithinBudget, buildDegradedLiveAssembleResult, buildForkBoundedLiveFallback, clampMessagesToSerializedBudget, resolveDeferredAssemblyPressure } from "./assemble-fallback.js";
+import { resolveBootstrapMaxTokens, trimBootstrapMessagesToBudget } from "./bootstrap-budget.js";
+import { batchLooksLikeHeartbeatAckTurn, filterSyntheticHeartbeatMessages, isHeartbeatNoiseContent, isHeartbeatOkContent, turnLooksLikeHeartbeatTurn } from "./heartbeat-filter.js";
+import { appendUncoveredVolatileLiveInputsWithinBudget, isVolatileLiveInputMessage, messageContentCoveredBySummary, resolveProtectedFreshTailAssembledIndexes } from "./live-coverage.js";
+import { RAW_PAYLOAD_EXTERNALIZATION_REASON, buildMessageParts, extractMessageContent, extractStructuredText, filterPersistableMessages, hasPersistableMessageRole, hasReplayCriticalRawBlock, isLikelyInjectedDeliveryOnlyTranscript, isLikelyInjectedMetadataPreambleRecord, isOpenClawRuntimeContextLeak, serializeRawPayloadContent, toStoredMessage, type StoredMessage } from "./message-content.js";
+import { createBootstrapEntryHash, createBootstrapReplaySignature, isBootstrapReplayCandidateMessage, messageIdentity, readBootstrapMessageFromJsonLine } from "./message-signatures.js";
+import { PROMPT_RECALL_MAX_MESSAGES, PROMPT_RECALL_SEARCH_CANDIDATE_LIMIT, buildPromptRecallProjectionFingerprint, extractPromptRecallIdentifiers, extractPromptRecallSnippet, findPromptRecallIdentifierIndex, isPromptRecallEligibleRole, normalizePromptRecallCoverageText, normalizePromptRecallText, renderPromptRecallMessage } from "./prompt-recall.js";
+import { externalizedReplayMetadataMatches, extractPlainToolReplayTextsById, extractRawBlockIdsFromPartMetadata, extractRawBlockSignatureFromPartMetadata, extractRawIdsFromPartMetadata, listTranscriptToolResultEntryIdsByCallId } from "./replay-metadata.js";
+import { estimateSessionTokenCountForAfterTurn, extractRuntimePromptTokenCount } from "./token-accounting.js";
+import { asRecord, formatDurationMs, isMissingFileError, normalizeSessionFilePathForComparison, safeBoolean, safeString } from "./value-utils.js";
 
 type AgentMessage = Parameters<ContextEngine["ingest"]>[0]["message"];
-type RepairLogger = { warn: (message: string) => void };
-
 const LOSSLESS_AGENT_RUN_REQUIRED_HOST_CAPABILITIES: ContextEngineHostCapability[] = [
   "bootstrap",
   "assemble-before-prompt",
@@ -120,19 +92,6 @@ const LOSSLESS_AGENT_RUN_REQUIRED_HOST_CAPABILITIES: ContextEngineHostCapability
 const LOSSLESS_SUBAGENT_SPAWN_REQUIRED_HOST_CAPABILITIES: ContextEngineHostCapability[] = [
   "thread-bootstrap-projection",
 ];
-type AssemblePrefixSnapshot = {
-  serializedMessages: string[];
-  messageSummaries: string[];
-  fullHash: string;
-};
-
-type BootstrapImportObservation = {
-  importedMessages: number;
-  reason: string | null;
-  forkBounded: boolean;
-  observedAt: Date;
-};
-
 const MAX_PREVIOUS_ASSEMBLED_SNAPSHOTS = 100;
 const FORK_BOUNDED_BOOTSTRAP_REASON = "fork-bounded bootstrap import";
 const AMBIGUOUS_SESSION_KEY_RUNTIME_ROLLOVER_REASON =
@@ -296,526 +255,10 @@ function checkpointIsPastTranscriptEof(
   return checkpoint.lastProcessedOffset > fileSize || checkpoint.lastSeenSize > fileSize;
 }
 
-function getErrorCode(error: unknown): string | undefined {
-  if (!(error instanceof Error)) {
-    return undefined;
-  }
-  const { code } = error as NodeJS.ErrnoException;
-  return typeof code === "string" ? code : undefined;
-}
-
-function isMissingFileError(error: unknown): boolean {
-  const code = getErrorCode(error);
-  return code === "ENOENT" || code === "ENOTDIR";
-}
-
-function normalizeSessionFilePathForComparison(filePath: string): string {
-  const trimmed = filePath.trim();
-  return trimmed ? resolvePath(trimmed) : "";
-}
-
 const TRANSCRIPT_GC_BATCH_SIZE = 12;
 const AUTO_ROTATE_DATABASE_LOCK_TIMEOUT_MS = 30_000;
 
 // ── Helpers ──────────────────────────────────────────────────────────────────
-
-function toJson(value: unknown): string {
-  const encoded = JSON.stringify(value);
-  return typeof encoded === "string" ? encoded : "";
-}
-
-function hashSerializedMessages(messages: string[]): string {
-  return createHash("sha256").update(JSON.stringify(messages)).digest("hex").slice(0, 16);
-}
-
-function normalizeDebugTextSnippet(value: string, maxLength: number = 48): string {
-  const collapsed = value.replace(/\s+/g, " ").trim();
-  if (collapsed.length <= maxLength) {
-    return collapsed;
-  }
-  return `${collapsed.slice(0, Math.max(0, maxLength - 3))}...`;
-}
-
-function summarizeMessageContentShape(content: unknown): string {
-  if (Array.isArray(content)) {
-    const blockTypes = content
-      .map((item) => {
-        const record = asRecord(item);
-        if (record) {
-          return safeString(record.type) ?? "object";
-        }
-        return typeof item;
-      })
-      .slice(0, 4);
-    const typeSummary = blockTypes.length > 0 ? blockTypes.join(",") : "empty";
-    return `blocks=${content.length}:${typeSummary}`;
-  }
-  if (typeof content === "string") {
-    return "content=text";
-  }
-  if (content == null) {
-    return "content=empty";
-  }
-  if (typeof content === "object") {
-    return "content=object";
-  }
-  return `content=${typeof content}`;
-}
-
-function summarizeMessageForPrefixDebug(message: AgentMessage): string {
-  const serialized = JSON.stringify(message);
-  const topLevel = message as Record<string, unknown>;
-  const role = safeString(topLevel.role) ?? "unknown";
-  const summaryParts = [role, summarizeMessageContentShape(topLevel.content)];
-  const toolCallId = extractTranscriptToolCallId(message);
-  if (toolCallId) {
-    summaryParts.push(`tool=${toolCallId}`);
-  }
-  const toolName =
-    safeString(topLevel.toolName) ??
-    safeString(topLevel.tool_name) ??
-    (Array.isArray(topLevel.content)
-      ? topLevel.content
-          .map((item) => asRecord(item))
-          .map((record) => safeString(record?.name))
-          .find((name) => typeof name === "string")
-      : undefined);
-  if (toolName) {
-    summaryParts.push(`name=${toolName}`);
-  }
-  const text = extractStructuredText(topLevel.content);
-  if (typeof text === "string" && text.trim().length > 0) {
-    summaryParts.push(`text=${toJson(normalizeDebugTextSnippet(text))}`);
-  }
-  summaryParts.push(
-    `hash=${createHash("sha256").update(serialized).digest("hex").slice(0, 8)}`,
-  );
-  return summaryParts.join("|");
-}
-
-function describeAssembledPrefixChange(
-  previous: AssemblePrefixSnapshot | undefined,
-  messages: AgentMessage[],
-): {
-  currentSnapshot: AssemblePrefixSnapshot;
-  previousCount: number;
-  commonPrefixCount: number;
-  commonPrefixHash: string;
-  previousWasPrefix: boolean;
-  firstDivergenceIndex: number;
-  previousDivergenceMessage: string;
-  currentDivergenceMessage: string;
-} {
-  const serializedMessages = messages.map((message) => JSON.stringify(message));
-  const messageSummaries = messages.map((message) => summarizeMessageForPrefixDebug(message));
-  const currentSnapshot = {
-    serializedMessages,
-    messageSummaries,
-    fullHash: hashSerializedMessages(serializedMessages),
-  };
-
-  if (!previous) {
-    return {
-      currentSnapshot,
-      previousCount: 0,
-      commonPrefixCount: 0,
-      commonPrefixHash: hashSerializedMessages([]),
-      previousWasPrefix: true,
-      firstDivergenceIndex: -1,
-      previousDivergenceMessage: "none",
-      currentDivergenceMessage: "none",
-    };
-  }
-
-  const limit = Math.min(previous.serializedMessages.length, serializedMessages.length);
-  let commonPrefixCount = 0;
-  while (
-    commonPrefixCount < limit &&
-    previous.serializedMessages[commonPrefixCount] === serializedMessages[commonPrefixCount]
-  ) {
-    commonPrefixCount++;
-  }
-
-  const previousWasPrefix = commonPrefixCount === previous.serializedMessages.length;
-  return {
-    currentSnapshot,
-    previousCount: previous.serializedMessages.length,
-    commonPrefixCount,
-    commonPrefixHash: hashSerializedMessages(serializedMessages.slice(0, commonPrefixCount)),
-    previousWasPrefix,
-    firstDivergenceIndex: previousWasPrefix ? -1 : commonPrefixCount,
-    previousDivergenceMessage: previousWasPrefix
-      ? "none"
-      : (previous.messageSummaries[commonPrefixCount] ?? "(end)"),
-    currentDivergenceMessage: previousWasPrefix
-      ? "none"
-      : (currentSnapshot.messageSummaries[commonPrefixCount] ?? "(end)"),
-  };
-}
-
-function shouldLogOverflowDiagnostics(params: {
-  diagnostics: AssemblyOverflowDiagnostics;
-  assembledTokens: number;
-  liveContextTokens: number;
-}): boolean {
-  const budget = Math.max(1, params.diagnostics.tokenBudget);
-  return (
-    params.diagnostics.totalContextTokens > budget ||
-    params.assembledTokens >= Math.floor(budget * 0.9) ||
-    params.liveContextTokens >= Math.floor(budget * 0.9) ||
-    params.diagnostics.duplicateRefClusters.length > 0 ||
-    params.diagnostics.duplicateMessageClusters.length > 0
-  );
-}
-
-function formatOverflowDiagnosticsForLog(params: {
-  diagnostics: AssemblyOverflowDiagnostics;
-  recentBootstrapImport?: BootstrapImportObservation;
-}): string {
-  const recent = params.recentBootstrapImport;
-  return JSON.stringify({
-    ...params.diagnostics,
-    recentBootstrapImportCount: recent?.importedMessages ?? null,
-    recentBootstrapImportReason: recent?.reason ?? null,
-  });
-}
-
-function safeString(value: unknown): string | undefined {
-  return typeof value === "string" ? value : undefined;
-}
-
-function extractRawIdsFromPartMetadata(metadata: string | null | undefined): string[] {
-  if (!metadata) {
-    return [];
-  }
-
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(metadata);
-  } catch {
-    return [];
-  }
-
-  const record = asRecord(parsed);
-  const raw = asRecord(record?.raw);
-
-  // Replay IDs can be preserved either inside the raw transcript block or
-  // as top-level metadata for string-content tool messages.
-  return [
-    safeString(raw?.id),
-    safeString(raw?.call_id),
-    safeString(raw?.toolCallId),
-    safeString(raw?.tool_call_id),
-    safeString(raw?.toolUseId),
-    safeString(raw?.tool_use_id),
-    safeString(record?.id),
-    safeString(record?.call_id),
-    safeString(record?.toolCallId),
-    safeString(record?.tool_call_id),
-    safeString(record?.toolUseId),
-    safeString(record?.tool_use_id),
-  ].filter((value): value is string => typeof value === "string" && value.length > 0);
-}
-
-function extractRawBlockIdsFromPartMetadata(metadata: string | null | undefined): string[] {
-  if (!metadata) {
-    return [];
-  }
-
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(metadata);
-  } catch {
-    return [];
-  }
-
-  const raw = asRecord(asRecord(parsed)?.raw);
-  return [
-    safeString(raw?.id),
-    safeString(raw?.call_id),
-    safeString(raw?.toolCallId),
-    safeString(raw?.tool_call_id),
-    safeString(raw?.toolUseId),
-    safeString(raw?.tool_use_id),
-  ].filter((value): value is string => typeof value === "string" && value.length > 0);
-}
-
-function extractRawBlockSignatureFromPartMetadata(metadata: string | null | undefined): string | null {
-  if (!metadata) {
-    return null;
-  }
-
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(metadata);
-  } catch {
-    return null;
-  }
-
-  const raw = asRecord(asRecord(parsed)?.raw);
-  return raw ? toJson(raw) : null;
-}
-
-function extractPlainToolReplayTextsById(message: AgentMessage): Map<string, string> {
-  const textsById = new Map<string, string>();
-  const duplicateIds = new Set<string>();
-  const addText = (replayId: string, text: string): void => {
-    if (duplicateIds.has(replayId)) {
-      return;
-    }
-    if (textsById.has(replayId)) {
-      textsById.delete(replayId);
-      duplicateIds.add(replayId);
-      return;
-    }
-    textsById.set(replayId, text);
-  };
-  if (
-    (message.role !== "toolResult" && message.role !== "tool") ||
-    !("content" in message)
-  ) {
-    return textsById;
-  }
-  const topLevel = message as unknown as Record<string, unknown>;
-  const topLevelToolCallId =
-    safeString(topLevel.toolCallId) ??
-    safeString(topLevel.tool_call_id) ??
-    safeString(topLevel.toolUseId) ??
-    safeString(topLevel.tool_use_id) ??
-    safeString(topLevel.call_id) ??
-    safeString(topLevel.id);
-  if (typeof message.content === "string") {
-    if (topLevelToolCallId) {
-      addText(topLevelToolCallId, message.content);
-    }
-    return textsById;
-  }
-  if (!Array.isArray(message.content)) {
-    return textsById;
-  }
-
-  for (const item of message.content) {
-    const record = asRecord(item);
-    if (!record) {
-      continue;
-    }
-    const rawType = safeString(record.type);
-    const replayId =
-      safeString(record.tool_use_id) ??
-      safeString(record.toolUseId) ??
-      safeString(record.tool_call_id) ??
-      safeString(record.toolCallId) ??
-      safeString(record.call_id) ??
-      safeString(record.id) ??
-      (message.content.length === 1 ? topLevelToolCallId : undefined);
-    if (!replayId) {
-      continue;
-    }
-
-    if (record.type === "text") {
-      const text = safeString(record.text);
-      if (text !== undefined) {
-        addText(replayId, text);
-      }
-      continue;
-    }
-    if (
-      rawType !== "tool_result" &&
-      rawType !== "toolResult" &&
-      rawType !== "function_call_output"
-    ) {
-      continue;
-    }
-    const textSource =
-      record.output !== undefined
-        ? record.output
-        : record.content !== undefined
-          ? record.content
-          : record;
-    const text = extractStructuredText(textSource);
-    if (text !== undefined) {
-      addText(replayId, text);
-    }
-  }
-  return textsById;
-}
-
-function stripExternalizedReplayMetadata(record: Record<string, unknown>): Record<string, unknown> {
-  const stripped = { ...record };
-  delete stripped.raw;
-  delete stripped.output;
-  delete stripped.content;
-  delete stripped.text;
-  delete stripped.externalizedFileId;
-  delete stripped.originalByteSize;
-  delete stripped.toolOutputExternalized;
-  delete stripped.externalizationReason;
-  delete stripped.rawType;
-  return stripped;
-}
-
-function canonicalizeReplayRawMetadata(record: Record<string, unknown>): Record<string, unknown> {
-  const canonical = stripExternalizedReplayMetadata(record);
-  const rawType = safeString(canonical.type);
-  if (rawType === "toolResult") {
-    canonical.type = "tool_result";
-  }
-
-  const replayId =
-    safeString(canonical.tool_use_id) ??
-    safeString(canonical.toolUseId) ??
-    safeString(canonical.tool_call_id) ??
-    safeString(canonical.toolCallId) ??
-    safeString(canonical.call_id) ??
-    safeString(canonical.id);
-  delete canonical.tool_use_id;
-  delete canonical.toolUseId;
-  delete canonical.tool_call_id;
-  delete canonical.toolCallId;
-  delete canonical.call_id;
-  delete canonical.id;
-  if (replayId) {
-    canonical[canonical.type === "function_call_output" ? "call_id" : "tool_use_id"] = replayId;
-  }
-
-  const isError = canonical.isError ?? canonical.is_error;
-  delete canonical.isError;
-  delete canonical.is_error;
-  if (typeof isError === "boolean") {
-    canonical.isError = isError;
-  }
-
-  return canonical;
-}
-
-function pickTopLevelReplayMetadata(record: Record<string, unknown>): Record<string, unknown> {
-  return {
-    originalRole: record.originalRole,
-    toolCallId: record.toolCallId,
-    toolName: record.toolName,
-    isError: record.isError,
-  };
-}
-
-function externalizedReplayMetadataMatches(
-  persistedMetadata: string | null,
-  incomingMetadata: string | null | undefined,
-): boolean {
-  let persistedParsed: unknown;
-  let incomingParsed: unknown;
-  try {
-    persistedParsed = persistedMetadata ? JSON.parse(persistedMetadata) : undefined;
-    incomingParsed = incomingMetadata ? JSON.parse(incomingMetadata) : undefined;
-  } catch {
-    return false;
-  }
-
-  const persistedRecord = asRecord(persistedParsed);
-  const incomingRecord = asRecord(incomingParsed);
-  if (!persistedRecord || !incomingRecord) {
-    return false;
-  }
-  const incomingRaw = asRecord(incomingRecord.raw);
-  if (!incomingRaw) {
-    return toJson(pickTopLevelReplayMetadata(persistedRecord)) ===
-      toJson(pickTopLevelReplayMetadata(incomingRecord));
-  }
-  if (
-    toJson(stripExternalizedReplayMetadata(persistedRecord)) !==
-    toJson(stripExternalizedReplayMetadata(incomingRecord))
-  ) {
-    return false;
-  }
-
-  const persistedRaw = asRecord(persistedRecord.raw);
-  return !!persistedRaw &&
-    toJson(canonicalizeReplayRawMetadata(persistedRaw)) ===
-      toJson(canonicalizeReplayRawMetadata(incomingRaw));
-}
-
-function formatDurationMs(durationMs: number): string {
-  return `${durationMs}ms`;
-}
-
-function asRecord(value: unknown): Record<string, unknown> | undefined {
-  return value && typeof value === "object" && !Array.isArray(value)
-    ? (value as Record<string, unknown>)
-    : undefined;
-}
-
-function safeBoolean(value: unknown): boolean | undefined {
-  return typeof value === "boolean" ? value : undefined;
-}
-
-function extractTranscriptToolCallId(message: AgentMessage): string | undefined {
-  const topLevel = message as Record<string, unknown>;
-  const direct =
-    safeString(topLevel.toolCallId) ??
-    safeString(topLevel.tool_call_id) ??
-    safeString(topLevel.toolUseId) ??
-    safeString(topLevel.tool_use_id) ??
-    safeString(topLevel.call_id) ??
-    safeString(topLevel.id);
-  if (direct) {
-    return direct;
-  }
-
-  if (!Array.isArray(topLevel.content)) {
-    return undefined;
-  }
-
-  for (const item of topLevel.content) {
-    const record = asRecord(item);
-    if (!record) {
-      continue;
-    }
-    const nested =
-      safeString(record.toolCallId) ??
-      safeString(record.tool_call_id) ??
-      safeString(record.toolUseId) ??
-      safeString(record.tool_use_id) ??
-      safeString(record.call_id) ??
-      safeString(record.id);
-    if (nested) {
-      return nested;
-    }
-  }
-
-  return undefined;
-}
-
-async function listTranscriptToolResultEntryIdsByCallId(
-  sessionFile: string,
-): Promise<Map<string, string>> {
-  const leafPathMessages = await readLeafPathMessages(sessionFile);
-  const entryIdsByCallId = new Map<string, string>();
-  const duplicateCallIds = new Set<string>();
-
-  for (const message of leafPathMessages) {
-    if (message.role !== "toolResult") {
-      continue;
-    }
-    const entryId = getTranscriptEntryId(message);
-    if (!entryId) {
-      continue;
-    }
-    const toolCallId = extractTranscriptToolCallId(message);
-    if (!toolCallId) {
-      continue;
-    }
-    if (entryIdsByCallId.has(toolCallId)) {
-      duplicateCallIds.add(toolCallId);
-      continue;
-    }
-    entryIdsByCallId.set(toolCallId, entryId);
-  }
-
-  for (const duplicateCallId of duplicateCallIds) {
-    entryIdsByCallId.delete(duplicateCallId);
-  }
-
-  return entryIdsByCallId;
-}
 
 function isRotatePreservedEntryType(type: string): boolean {
   return (
@@ -834,1115 +277,6 @@ function normalizeRotateTailMessageCount(value: number, branchMessageCount: numb
     return 1;
   }
   return Math.max(1, Math.min(branchMessageCount, Math.floor(value)));
-}
-
-function appendTextValue(value: unknown, out: string[]): void {
-  if (typeof value === "string") {
-    out.push(value);
-    return;
-  }
-  if (Array.isArray(value)) {
-    for (const entry of value) {
-      appendTextValue(entry, out);
-    }
-    return;
-  }
-  if (!value || typeof value !== "object") {
-    return;
-  }
-
-  const record = value as Record<string, unknown>;
-  appendTextValue(record.text, out);
-  appendTextValue(record.value, out);
-}
-
-const STRUCTURED_TEXT_FIELD_KEYS = ["text", "transcript", "transcription", "message", "summary"];
-const STRUCTURED_ARRAY_FIELD_KEYS = [
-  "segments",
-  "utterances",
-  "paragraphs",
-  "alternatives",
-  "words",
-  "items",
-  "results",
-];
-const STRUCTURED_NESTED_FIELD_KEYS = ["content", "output", "result", "payload", "data", "value"];
-const MAX_STRUCTURED_TEXT_DEPTH = 6;
-const TOOL_CALL_RAW_TYPES: ReadonlySet<string> = new Set([
-  "tool_use",
-  "toolUse",
-  "tool-use",
-  "toolCall",
-  "tool_call",
-  "functionCall",
-  "function_call",
-]);
-const TOOL_RESULT_RAW_TYPES: ReadonlySet<string> = new Set([
-  "function_call_output",
-  "tool_result",
-  "toolResult",
-  "tool_use_result",
-]);
-const TOOL_RAW_TYPES: ReadonlySet<string> = new Set([
-  ...TOOL_CALL_RAW_TYPES,
-  ...TOOL_RESULT_RAW_TYPES,
-]);
-const REASONING_RAW_TYPES: ReadonlySet<string> = new Set([
-  "thinking",
-  "redacted_thinking",
-  "reasoning",
-]);
-const REPLAY_CRITICAL_RAW_TYPES: ReadonlySet<string> = new Set([
-  ...TOOL_RAW_TYPES,
-  ...REASONING_RAW_TYPES,
-]);
-const RAW_PAYLOAD_EXTERNALIZATION_REASON = "large_raw_message";
-
-function looksLikeJsonPayload(value: string): boolean {
-  if (typeof value !== "string") return false;
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return false;
-  }
-  return (
-    (trimmed.startsWith("{") && trimmed.endsWith("}")) ||
-    (trimmed.startsWith("[") && trimmed.endsWith("]"))
-  );
-}
-
-function extractStructuredText(value: unknown, depth: number = 0): string | undefined {
-  if (value == null || depth > MAX_STRUCTURED_TEXT_DEPTH) {
-    return undefined;
-  }
-  if (typeof value === "string") {
-    if (looksLikeJsonPayload(value)) {
-      try {
-        const parsed = JSON.parse(value.trim());
-        const parsedText = extractStructuredText(parsed, depth + 1);
-        if (typeof parsedText === "string" && parsedText.length > 0) {
-          return parsedText;
-        }
-      } catch {
-        // Fall through to returning the original string when parsing fails.
-      }
-    }
-    return value;
-  }
-  if (Array.isArray(value)) {
-    const texts: string[] = [];
-    for (const entry of value) {
-      const text = extractStructuredText(entry, depth + 1);
-      if (typeof text === "string" && text.trim().length > 0) {
-        texts.push(text);
-      }
-    }
-    return texts.length > 0 ? texts.join("\n") : undefined;
-  }
-  if (typeof value !== "object") {
-    return undefined;
-  }
-
-  const record = value as Record<string, unknown>;
-
-  if (typeof record.type === "string" && REASONING_RAW_TYPES.has(record.type)) {
-    return undefined;
-  }
-
-  // Skip tool call/result objects — their structured data belongs in the parts table, not content
-  if (typeof record.type === "string" && TOOL_RAW_TYPES.has(record.type)) {
-    if (safeBoolean(record.toolOutputExternalized)) {
-      const externalizedText =
-        extractStructuredText(record.output, depth + 1) ??
-        extractStructuredText(record.content, depth + 1) ??
-        extractStructuredText(record.result, depth + 1);
-      if (typeof externalizedText === "string" && externalizedText.trim().length > 0) {
-        return externalizedText;
-      }
-    }
-    return undefined;
-  }
-
-  for (const key of STRUCTURED_TEXT_FIELD_KEYS) {
-    const candidate = record[key];
-    if (typeof candidate === "string" && candidate.trim().length > 0) {
-      return candidate;
-    }
-  }
-
-  for (const key of STRUCTURED_ARRAY_FIELD_KEYS) {
-    const candidate = record[key];
-    if (Array.isArray(candidate)) {
-      const texts: string[] = [];
-      for (const entry of candidate) {
-        const text = extractStructuredText(entry, depth + 1);
-        if (typeof text === "string" && text.trim().length > 0) {
-          texts.push(text);
-        }
-      }
-      if (texts.length > 0) {
-        return texts.join("\n");
-      }
-    }
-  }
-
-  for (const key of STRUCTURED_NESTED_FIELD_KEYS) {
-    const nested = record[key];
-    const nestedText = extractStructuredText(nested, depth + 1);
-    if (typeof nestedText === "string" && nestedText.trim().length > 0) {
-      return nestedText;
-    }
-  }
-
-  return undefined;
-}
-
-function extractReasoningText(record: Record<string, unknown>): string | undefined {
-  const chunks: string[] = [];
-  appendTextValue(record.summary, chunks);
-  if (chunks.length === 0) {
-    return undefined;
-  }
-
-  const normalized = chunks
-    .map((chunk) => chunk.trim())
-    .filter((chunk, idx, arr) => chunk.length > 0 && arr.indexOf(chunk) === idx);
-  return normalized.length > 0 ? normalized.join("\n") : undefined;
-}
-
-/** Return true when a raw block should remain structurally replayable. */
-function hasReplayCriticalRawBlock(value: unknown): boolean {
-  if (!value || typeof value !== "object") {
-    return false;
-  }
-  if (Array.isArray(value)) {
-    return value.some((entry) => hasReplayCriticalRawBlock(entry));
-  }
-
-  const record = value as Record<string, unknown>;
-  const rawType = safeString(record.type) ?? safeString(record.rawType);
-  if (rawType && REPLAY_CRITICAL_RAW_TYPES.has(rawType)) {
-    return true;
-  }
-
-  for (const key of STRUCTURED_NESTED_FIELD_KEYS) {
-    if (hasReplayCriticalRawBlock(record[key])) {
-      return true;
-    }
-  }
-  for (const key of STRUCTURED_ARRAY_FIELD_KEYS) {
-    if (hasReplayCriticalRawBlock(record[key])) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-/** Serialize the original message content that backs a generic raw-payload reference. */
-function serializeRawPayloadContent(message: AgentMessage, fallbackContent: string): {
-  content: string;
-  mimeType: string;
-} | null {
-  if (!("content" in message)) {
-    return null;
-  }
-  if (typeof message.content === "string") {
-    return {
-      content: message.content,
-      mimeType: "text/plain",
-    };
-  }
-
-  const serialized = JSON.stringify(message.content);
-  if (typeof serialized !== "string") {
-    return null;
-  }
-  return {
-    content: serialized || fallbackContent,
-    mimeType: "application/json",
-  };
-}
-
-function normalizeUnknownBlock(value: unknown): {
-  type: string;
-  text?: string;
-  metadata: Record<string, unknown>;
-} {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return {
-      type: "agent",
-      metadata: { raw: value },
-    };
-  }
-
-  const record = value as Record<string, unknown>;
-  const rawType = safeString(record.type);
-  return {
-    type: rawType ?? "agent",
-    text:
-      safeString(record.text) ??
-      safeString(record.thinking) ??
-      ((rawType === "reasoning" || rawType === "thinking")
-        ? extractReasoningText(record)
-        : undefined),
-    metadata: { raw: record },
-  };
-}
-
-function extractTopLevelReasoningContent(
-  role: string,
-  topLevel: Record<string, unknown>,
-): { field: "reasoning_content"; content: string } | null {
-  if (role !== "assistant") {
-    return null;
-  }
-  const content = safeString(topLevel.reasoning_content);
-  return content && content.trim().length > 0
-    ? { field: "reasoning_content", content }
-    : null;
-}
-
-function topLevelReasoningMetadata(
-  reasoning: { field: "reasoning_content"; content: string } | null,
-  only = false,
-): Record<string, unknown> {
-  if (!reasoning) {
-    return {};
-  }
-  return {
-    topLevelReasoningField: reasoning.field,
-    topLevelReasoningContent: reasoning.content,
-    topLevelReasoningOnly: only || undefined,
-  };
-}
-
-function toPartType(type: string): MessagePartType {
-  switch (type) {
-    case "text":
-      return "text";
-    case "thinking":
-    case "redacted_thinking":
-    case "reasoning":
-      return "reasoning";
-    case "tool_use":
-    case "toolUse":
-    case "tool-use":
-    case "toolCall":
-    case "functionCall":
-    case "function_call":
-    case "function_call_output":
-    case "tool_result":
-    case "toolResult":
-    case "tool":
-      return "tool";
-    case "patch":
-      return "patch";
-    case "file":
-    case "image":
-      return "file";
-    case "subtask":
-      return "subtask";
-    case "compaction":
-      return "compaction";
-    case "step_start":
-    case "step-start":
-      return "step_start";
-    case "step_finish":
-    case "step-finish":
-      return "step_finish";
-    case "snapshot":
-      return "snapshot";
-    case "retry":
-      return "retry";
-    case "agent":
-      return "agent";
-    default:
-      return "agent";
-  }
-}
-
-/**
- * Convert AgentMessage content into plain text for DB storage.
- *
- * For content block arrays we keep only text blocks to avoid persisting raw
- * JSON syntax that can later pollute assembled model context.
- */
-function extractMessageContent(content: unknown): string {
-  const extracted = extractStructuredText(content);
-  if (typeof extracted === "string") {
-    return extracted;
-  }
-  if (content == null) {
-    return "";
-  }
-  if (Array.isArray(content) && content.length === 0) {
-    return "";
-  }
-  // If content is an array of only tool call/result/reasoning objects, store as empty
-  // (structured data is preserved in the message parts table)
-  if (Array.isArray(content) && content.length > 0 && content.every(
-    (item) => typeof item === "object" && item !== null && !Array.isArray(item) &&
-      typeof (item as Record<string, unknown>).type === "string" &&
-      (
-        TOOL_RAW_TYPES.has((item as Record<string, unknown>).type as string) ||
-        REASONING_RAW_TYPES.has((item as Record<string, unknown>).type as string)
-      )
-  )) {
-    return "";
-  }
-
-  const serialized = JSON.stringify(content);
-  return typeof serialized === "string" ? serialized : "";
-}
-
-function toRuntimeRoleForTokenEstimate(role: string): "user" | "assistant" | "toolResult" {
-  if (role === "tool" || role === "toolResult") {
-    return "toolResult";
-  }
-  if (role === "user" || role === "system") {
-    return "user";
-  }
-  return "assistant";
-}
-
-function isTextBlock(value: unknown): value is { type: "text"; text: string } {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return false;
-  }
-  const record = value as Record<string, unknown>;
-  return record.type === "text" && typeof record.text === "string";
-}
-
-function toSyntheticMessagePartRecord(
-  part: CreateMessagePartInput,
-  messageId: number,
-): MessagePartRecord {
-  return {
-    partId: `estimate-part-${part.ordinal}`,
-    messageId,
-    sessionId: part.sessionId,
-    partType: part.partType,
-    ordinal: part.ordinal,
-    textContent: part.textContent ?? null,
-    toolCallId: part.toolCallId ?? null,
-    toolName: part.toolName ?? null,
-    toolInput: part.toolInput ?? null,
-    toolOutput: part.toolOutput ?? null,
-    metadata: part.metadata ?? null,
-  };
-}
-
-function normalizeMessageContentForStorage(params: {
-  message: AgentMessage;
-  fallbackContent: string;
-}): unknown {
-  const { message, fallbackContent } = params;
-  if (!("content" in message)) {
-    return fallbackContent;
-  }
-
-  const role = toRuntimeRoleForTokenEstimate(message.role);
-  const parts = buildMessageParts({
-    sessionId: "storage-estimate",
-    message,
-    fallbackContent,
-  }).map((part) => toSyntheticMessagePartRecord(part, 0));
-
-  if (parts.length === 0) {
-    if (role === "assistant") {
-      return fallbackContent ? [{ type: "text", text: fallbackContent }] : [];
-    }
-    if (role === "toolResult") {
-      return [{ type: "text", text: fallbackContent }];
-    }
-    return fallbackContent;
-  }
-
-  const blocks = parts.map(blockFromPart);
-  if (role === "user" && blocks.length === 1 && isTextBlock(blocks[0])) {
-    return blocks[0].text;
-  }
-  return blocks;
-}
-
-/**
- * Estimate token usage for the content shape that the assembler will emit.
- *
- * LCM stores a plain-text fallback copy in messages.content, but message_parts
- * can rehydrate larger structured/raw blocks. This estimator mirrors the
- * rehydrated shape so compaction decisions use realistic token totals.
- */
-function estimateContentTokensForRole(params: {
-  role: "user" | "assistant" | "toolResult";
-  content: unknown;
-  fallbackContent: string;
-}): number {
-  const { role, content, fallbackContent } = params;
-
-  if (typeof content === "string") {
-    return estimateTokens(content);
-  }
-
-  if (Array.isArray(content)) {
-    if (content.length === 0) {
-      return estimateTokens(fallbackContent);
-    }
-
-    if (role === "user" && content.length === 1 && isTextBlock(content[0])) {
-      return estimateTokens(content[0].text);
-    }
-
-    const serialized = JSON.stringify(content);
-    return estimateTokens(typeof serialized === "string" ? serialized : "");
-  }
-
-  if (content && typeof content === "object") {
-    if (role === "user" && isTextBlock(content)) {
-      return estimateTokens(content.text);
-    }
-
-    const serialized = JSON.stringify([content]);
-    return estimateTokens(typeof serialized === "string" ? serialized : "");
-  }
-
-  return estimateTokens(fallbackContent);
-}
-
-function buildMessageParts(params: {
-  sessionId: string;
-  message: AgentMessage;
-  fallbackContent: string;
-}): import("./store/conversation-store.js").CreateMessagePartInput[] {
-  const { sessionId, message, fallbackContent } = params;
-  const role = typeof message.role === "string" ? message.role : "unknown";
-  const topLevel = message as unknown as Record<string, unknown>;
-  const topLevelToolCallId =
-    safeString(topLevel.toolCallId) ??
-    safeString(topLevel.tool_call_id) ??
-    safeString(topLevel.toolUseId) ??
-    safeString(topLevel.tool_use_id) ??
-    safeString(topLevel.call_id) ??
-    safeString(topLevel.id);
-  const topLevelToolName =
-    safeString(topLevel.toolName) ??
-    safeString(topLevel.tool_name);
-  const topLevelIsError =
-    safeBoolean(topLevel.isError) ??
-    safeBoolean(topLevel.is_error);
-  const topLevelReasoning = extractTopLevelReasoningContent(role, topLevel);
-  const rawPayloadExternalized = safeBoolean(topLevel.rawPayloadExternalized);
-  const externalizedFileId = safeString(topLevel.externalizedFileId);
-  const originalByteSize =
-    typeof topLevel.originalByteSize === "number"
-      ? topLevel.originalByteSize
-      : undefined;
-  const externalizationReason = safeString(topLevel.externalizationReason);
-
-  // BashExecutionMessage: preserve a synthetic text part so output is round-trippable.
-  if (!("content" in message) && "command" in message && "output" in message) {
-    return [
-      {
-        sessionId,
-        partType: "text",
-        ordinal: 0,
-        textContent: fallbackContent,
-        metadata: toJson({
-          originalRole: role,
-          source: "bash-exec",
-          command: safeString((message as { command?: unknown }).command),
-        }),
-      },
-    ];
-  }
-
-  if (!("content" in message)) {
-    return [
-      {
-        sessionId,
-        partType: "agent",
-        ordinal: 0,
-        textContent: fallbackContent || null,
-        metadata: toJson({
-          originalRole: role,
-          source: "unknown-message-shape",
-          raw: message,
-        }),
-      },
-    ];
-  }
-
-  if (typeof message.content === "string") {
-    return [
-      {
-        sessionId,
-        partType: "text",
-        ordinal: 0,
-        textContent: message.content,
-        metadata: toJson({
-          originalRole: role,
-          toolCallId: topLevelToolCallId,
-          toolName: topLevelToolName,
-          isError: topLevelIsError,
-          ...topLevelReasoningMetadata(topLevelReasoning),
-          rawPayloadExternalized: rawPayloadExternalized || undefined,
-          externalizedFileId,
-          originalByteSize,
-          externalizationReason,
-        }),
-      },
-    ];
-  }
-
-  if (!Array.isArray(message.content)) {
-    return [
-      {
-        sessionId,
-        partType: "agent",
-        ordinal: 0,
-        textContent: fallbackContent || null,
-        metadata: toJson({
-          originalRole: role,
-          source: "non-array-content",
-          raw: message.content,
-          ...topLevelReasoningMetadata(topLevelReasoning),
-        }),
-      },
-    ];
-  }
-
-  const parts: CreateMessagePartInput[] = [];
-  if (message.content.length === 0 && topLevelReasoning) {
-    parts.push({
-      sessionId,
-      partType: "reasoning",
-      ordinal: 0,
-      textContent: null,
-      metadata: toJson({
-        originalRole: role,
-        rawType: topLevelReasoning.field,
-        ...topLevelReasoningMetadata(topLevelReasoning, true),
-      }),
-    });
-  }
-  for (let ordinal = 0; ordinal < message.content.length; ordinal++) {
-    const block = normalizeUnknownBlock(message.content[ordinal]);
-    const metadataRecord = block.metadata.raw as Record<string, unknown> | undefined;
-    const rawBlockType = safeString(metadataRecord?.rawType) ?? block.type;
-    const partType = toPartType(rawBlockType);
-    const rawBlock =
-      metadataRecord && rawBlockType !== block.type
-        ? {
-            ...metadataRecord,
-            type: rawBlockType,
-          }
-        : (metadataRecord ?? message.content[ordinal]);
-    const toolCallId =
-      safeString(metadataRecord?.toolCallId) ??
-      safeString(metadataRecord?.tool_call_id) ??
-      safeString(metadataRecord?.toolUseId) ??
-      safeString(metadataRecord?.tool_use_id) ??
-      safeString(metadataRecord?.call_id) ??
-      (partType === "tool" ? safeString(metadataRecord?.id) : undefined) ??
-      topLevelToolCallId;
-
-    parts.push({
-      sessionId,
-      partType,
-      ordinal,
-      textContent: block.text ?? null,
-      toolCallId,
-      toolName:
-        safeString(metadataRecord?.name) ??
-        safeString(metadataRecord?.toolName) ??
-        safeString(metadataRecord?.tool_name) ??
-        topLevelToolName,
-      toolInput:
-        metadataRecord?.input !== undefined
-          ? toJson(metadataRecord.input)
-          : metadataRecord?.arguments !== undefined
-            ? toJson(metadataRecord.arguments)
-          : metadataRecord?.toolInput !== undefined
-            ? toJson(metadataRecord.toolInput)
-            : (safeString(metadataRecord?.tool_input) ?? null),
-      toolOutput:
-        metadataRecord?.output !== undefined
-          ? toJson(metadataRecord.output)
-          : metadataRecord?.toolOutput !== undefined
-            ? toJson(metadataRecord.toolOutput)
-            : (safeString(metadataRecord?.tool_output) ?? null),
-      metadata: toJson({
-        originalRole: role,
-        toolCallId: topLevelToolCallId,
-        toolName: topLevelToolName,
-        isError: topLevelIsError,
-        ...(ordinal === 0 ? topLevelReasoningMetadata(topLevelReasoning) : {}),
-        externalizedFileId: safeString(metadataRecord?.externalizedFileId),
-        originalByteSize:
-          typeof metadataRecord?.originalByteSize === "number"
-            ? metadataRecord.originalByteSize
-            : undefined,
-        toolOutputExternalized: safeBoolean(metadataRecord?.toolOutputExternalized),
-        externalizationReason: safeString(metadataRecord?.externalizationReason),
-        rawType: rawBlockType,
-        raw: rawBlock,
-      }),
-    });
-  }
-
-  return parts;
-}
-
-/**
- * Map AgentMessage role to the DB enum.
- *
- *   "user"      -> "user"
- *   "assistant" -> "assistant"
- *
- * AgentMessage only has user/assistant roles, but we keep the mapping
- * explicit for clarity and future-proofing.
- */
-function toDbRole(role: string): "user" | "assistant" | "system" | "tool" {
-  if (role === "tool" || role === "toolResult") {
-    return "tool";
-  }
-  if (role === "system") {
-    return "system";
-  }
-  if (role === "user") {
-    return "user";
-  }
-  if (role === "assistant") {
-    return "assistant";
-  }
-  // Direct callers should filter unknown roles before storage. Preserve the
-  // historical fallback for typed AgentMessage values that reach this helper.
-  return "assistant";
-}
-
-function hasPersistableMessageRole(message: AgentMessage): boolean {
-  const role = (message as { role?: unknown }).role;
-  return (
-    role === "user" ||
-    role === "assistant" ||
-    role === "system" ||
-    role === "tool" ||
-    role === "toolResult"
-  );
-}
-
-function filterPersistableMessages(messages: AgentMessage[]): AgentMessage[] {
-  return messages.filter(hasPersistableMessageRole);
-}
-
-type StoredMessage = {
-  role: "user" | "assistant" | "system" | "tool";
-  content: string;
-  tokenCount: number;
-};
-
-const PROMPT_RECALL_IDENTIFIER_PATTERN = /\b[A-Z][A-Z0-9]*(?:_[A-Z0-9]+)+\b/g;
-const PROMPT_RECALL_MAX_IDENTIFIERS = 4;
-const PROMPT_RECALL_MAX_MESSAGES = 4;
-const PROMPT_RECALL_MAX_MESSAGE_CHARS = 1200;
-const PROMPT_RECALL_SEARCH_LIMIT = PROMPT_RECALL_MAX_MESSAGES * 2;
-const PROMPT_RECALL_SEARCH_CANDIDATE_LIMIT = PROMPT_RECALL_SEARCH_LIMIT * 4;
-const DELIVERY_ONLY_TRANSCRIPT_MAX_MESSAGES = 4;
-const INJECTED_DELIVERY_TRANSCRIPT_PATTERN = /\b(?:delivery[-_\s]?mirror|config[-_\s]?audit)\b/i;
-const INJECTED_METADATA_PREAMBLE_PREFIX = "Conversation info (untrusted metadata)";
-const OPENCLAW_RUNTIME_CONTEXT_SENTINEL =
-  "OpenClaw runtime context for the immediately preceding user message. This context is runtime-generated, not user-author.";
-const PROMPT_RECALL_SENSITIVE_IDENTIFIER_PATTERN =
-  /(?:^|[^A-Za-z0-9])(?:ACCESS_?KEY|API_?KEY|AUTH|CREDENTIALS?|DEPLOY_?KEY|KEY|PASS(?:WORD)?|PRIVATE_?KEY|SECRET|TOKEN)(?=$|[^A-Za-z0-9])/i;
-const PROMPT_RECALL_SENSITIVE_VALUE_PATTERN =
-  /(?:-----BEGIN [A-Z ]*PRIVATE KEY-----|\bAKIA[0-9A-Z]{16}\b|\b(?:ghp|gho|ghu|ghs|ghr)_[A-Za-z0-9_]{10,}\b|\bgithub_pat_[A-Za-z0-9_]{20,}\b|\bxox[baprs]-[A-Za-z0-9-]{10,}\b|\b(?:sk|rk|pk)-[A-Za-z0-9_-]{10,}\b|\b(?:sk|rk|pk)_[A-Za-z0-9_]{10,}\b)/i;
-
-/**
- * Normalize AgentMessage variants into the storage shape used by LCM.
- */
-function toStoredMessage(message: AgentMessage): StoredMessage {
-  const content =
-    "content" in message
-      ? extractMessageContent(message.content)
-      : "output" in message
-        ? `$ ${String(message.command ?? "")}\n${String(message.output)}`
-        : "";
-  const runtimeRole = toRuntimeRoleForTokenEstimate(message.role);
-  const normalizedContent =
-    "content" in message
-      ? normalizeMessageContentForStorage({
-          message,
-          fallbackContent: content,
-        })
-      : content;
-  const tokenCount =
-    "content" in message
-      ? estimateContentTokensForRole({
-          role: runtimeRole,
-          content: normalizedContent,
-          fallbackContent: content,
-        })
-      : estimateTokens(content);
-  const topLevelReasoning = extractTopLevelReasoningContent(
-    typeof message.role === "string" ? message.role : "",
-    message as unknown as Record<string, unknown>,
-  );
-
-  return {
-    role: toDbRole(message.role),
-    content,
-    tokenCount: tokenCount + (topLevelReasoning ? estimateTokens(topLevelReasoning.content) : 0),
-  };
-}
-
-function isLikelyInjectedDeliveryMessage(message: AgentMessage): boolean {
-  const stored = toStoredMessage(message);
-  return stored.role === "system" && INJECTED_DELIVERY_TRANSCRIPT_PATTERN.test(stored.content);
-}
-
-function isOpenClawRuntimeContextLeak(stored: StoredMessage): boolean {
-  return (
-    stored.role === "assistant" &&
-    stored.content.trimStart().startsWith(OPENCLAW_RUNTIME_CONTEXT_SENTINEL)
-  );
-}
-
-function isLikelyInjectedDeliveryOnlyTranscript(messages: AgentMessage[]): boolean {
-  return (
-    messages.length > 0 &&
-    messages.length <= DELIVERY_ONLY_TRANSCRIPT_MAX_MESSAGES &&
-    messages.every(isLikelyInjectedDeliveryMessage)
-  );
-}
-
-function isLikelyInjectedMetadataPreambleRecord(message: {
-  role: string;
-  content: string;
-}): boolean {
-  return (
-    message.role === "user" &&
-    message.content.trimStart().startsWith(INJECTED_METADATA_PREAMBLE_PREFIX)
-  );
-}
-
-function escapeRegexLiteral(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function isPromptRecallSensitiveIdentifier(identifier: string): boolean {
-  return PROMPT_RECALL_SENSITIVE_IDENTIFIER_PATTERN.test(identifier);
-}
-
-function containsPromptRecallSensitiveMaterial(value: string): boolean {
-  return (
-    PROMPT_RECALL_SENSITIVE_IDENTIFIER_PATTERN.test(value) ||
-    PROMPT_RECALL_SENSITIVE_VALUE_PATTERN.test(value)
-  );
-}
-
-function findPromptRecallIdentifierIndex(content: string, identifier: string): number {
-  const match = new RegExp(
-    `(^|[^A-Za-z0-9_])${escapeRegexLiteral(identifier)}($|[^A-Za-z0-9_])`,
-  ).exec(content);
-  return match ? match.index + (match[1]?.length ?? 0) : -1;
-}
-
-function findPromptRecallLineStart(content: string, identifierIndex: number): number {
-  const searchStart = Math.max(0, identifierIndex - 1);
-  const previousLineBreak = Math.max(
-    content.lastIndexOf("\n", searchStart),
-    content.lastIndexOf("\r", searchStart),
-  );
-  return previousLineBreak >= 0 ? previousLineBreak + 1 : 0;
-}
-
-function findPromptRecallLineEnd(content: string, identifierIndex: number): number {
-  const nextLineFeed = content.indexOf("\n", identifierIndex);
-  const nextCarriageReturn = content.indexOf("\r", identifierIndex);
-  if (nextLineFeed < 0) {
-    return nextCarriageReturn >= 0 ? nextCarriageReturn : content.length;
-  }
-  if (nextCarriageReturn < 0) {
-    return nextLineFeed;
-  }
-  return Math.min(nextLineFeed, nextCarriageReturn);
-}
-
-function findPromptRecallSentenceStart(line: string, relativeIdentifierIndex: number): number {
-  let sentenceStart = 0;
-  for (const match of line.slice(0, relativeIdentifierIndex).matchAll(/[.!?](?:\s+|$)/g)) {
-    sentenceStart = (match.index ?? 0) + match[0].length;
-  }
-  return sentenceStart;
-}
-
-function findPromptRecallSentenceEnd(
-  line: string,
-  relativeIdentifierIndex: number,
-  identifierLength: number,
-): number {
-  const afterIdentifierStart = relativeIdentifierIndex + identifierLength;
-  const match = /[.!?](?:\s|$)/.exec(line.slice(afterIdentifierStart));
-  return match ? afterIdentifierStart + match.index + 1 : line.length;
-}
-
-function clipPromptRecallSnippet(snippet: string, identifier: string): string {
-  if (snippet.length <= PROMPT_RECALL_MAX_MESSAGE_CHARS) {
-    return snippet;
-  }
-  const identifierIndex = findPromptRecallIdentifierIndex(snippet, identifier);
-  if (identifierIndex < 0) {
-    return snippet.slice(0, PROMPT_RECALL_MAX_MESSAGE_CHARS);
-  }
-  const preferredContextBeforeIdentifier = Math.floor(PROMPT_RECALL_MAX_MESSAGE_CHARS * 0.75);
-  const start = Math.max(0, identifierIndex - preferredContextBeforeIdentifier);
-  const end = Math.min(snippet.length, start + PROMPT_RECALL_MAX_MESSAGE_CHARS);
-  return `${start > 0 ? "..." : ""}${snippet.slice(start, end)}${end < snippet.length ? "..." : ""}`;
-}
-
-function extractPromptRecallSnippet(content: string, identifier: string): string | null {
-  const identifierIndex = findPromptRecallIdentifierIndex(content, identifier);
-  if (identifierIndex < 0) {
-    return null;
-  }
-  const lineStart = findPromptRecallLineStart(content, identifierIndex);
-  const lineEnd = findPromptRecallLineEnd(content, identifierIndex);
-  const line = content.slice(lineStart, lineEnd);
-  const relativeIdentifierIndex = identifierIndex - lineStart;
-  const sentenceStart = findPromptRecallSentenceStart(line, relativeIdentifierIndex);
-  const sentenceEnd = findPromptRecallSentenceEnd(line, relativeIdentifierIndex, identifier.length);
-  const rawSnippet = clipPromptRecallSnippet(line.slice(sentenceStart, sentenceEnd), identifier);
-  if (containsPromptRecallSensitiveMaterial(rawSnippet)) {
-    return null;
-  }
-  const snippet = normalizePromptRecallText(rawSnippet);
-  return snippet.length > 0 ? snippet : null;
-}
-
-function isPromptRecallEligibleRole(role: StoredMessage["role"]): boolean {
-  return role === "user" || role === "assistant";
-}
-
-function extractPromptRecallIdentifiers(prompt?: string): string[] {
-  if (typeof prompt !== "string" || !prompt.trim()) {
-    return [];
-  }
-  return [...new Set(prompt.match(PROMPT_RECALL_IDENTIFIER_PATTERN) ?? [])]
-    .filter((identifier) => !isPromptRecallSensitiveIdentifier(identifier))
-    .slice(
-      0,
-      PROMPT_RECALL_MAX_IDENTIFIERS,
-    );
-}
-
-function renderPromptRecallMessage(params: {
-  identifier: string;
-  role: StoredMessage["role"];
-  content: string;
-}): string {
-  const singleLine = normalizePromptRecallText(params.content);
-  const clipped =
-    singleLine.length > PROMPT_RECALL_MAX_MESSAGE_CHARS
-      ? `${singleLine.slice(0, PROMPT_RECALL_MAX_MESSAGE_CHARS)}...`
-      : singleLine;
-  return `- ${params.role} matched ${params.identifier}: ${JSON.stringify(clipped)}`;
-}
-
-function normalizePromptRecallText(value: string): string {
-  return value.replace(/\s+/g, " ").trim();
-}
-
-function normalizePromptRecallCoverageText(value: string): string {
-  return normalizePromptRecallText(value).replace(/[.!?]$/, "");
-}
-
-function buildPromptRecallProjectionFingerprint(message: AgentMessage): string {
-  const content = "content" in message ? extractMessageContent(message.content) : JSON.stringify(message);
-  return [
-    "prompt-recall-v1",
-    createHash("sha256").update(content).digest("hex").slice(0, 32),
-  ].join(":");
-}
-
-function createBootstrapEntryHash(message: StoredMessage | null): string | null {
-  if (!message) {
-    return null;
-  }
-  return createHash("sha256")
-    .update(JSON.stringify({ role: message.role, content: message.content }))
-    .digest("hex");
-}
-
-/**
- * Estimate live transcript tokens at the model boundary.
- *
- * Uses full-message serialization so structured tool-call payloads count.
- * A text-block-only estimate undercounts tool-heavy transcripts by 2-3x,
- * which previously let over-budget prompts pass every live pressure check
- * while the host's LLM-boundary estimate rejected them.
- */
-function estimateSessionTokenCountForAfterTurn(messages: AgentMessage[]): number {
-  return estimateSerializedMessagesTokens(messages);
-}
-
-function normalizeNonNegativeInteger(value: unknown): number | undefined {
-  if (typeof value !== "number" || !Number.isFinite(value) || value < 0) {
-    return undefined;
-  }
-  return Math.floor(value);
-}
-
-function firstRuntimeTokenCount(record: Record<string, unknown> | null, keys: string[]): number | undefined {
-  if (!record) {
-    return undefined;
-  }
-  for (const key of keys) {
-    const count = normalizeNonNegativeInteger(record[key]);
-    if (count !== undefined) {
-      return count;
-    }
-  }
-  return undefined;
-}
-
-/**
- * Extract the runtime prompt token count from OpenClaw runtimeContext.
- *
- * OpenClaw derives this as: input + cacheRead + cacheWrite from the
- * normalizeUsage() result.  The runtimeContext carries it three ways:
- *   1. runtimeContext.currentTokenCount  — direct value (preferred)
- *   2. runtimeContext.usage             — {input, cacheRead, cacheWrite, ...}
- *   3. runtimeContext.promptCache.lastCallUsage — same normalized shape
- *
- * normalizeUsage() maps provider-specific fields (prompt_tokens, input_tokens,
- * cache_read, etc.) to the canonical {input, cacheRead, cacheWrite} shape,
- * so the lastCallUsage passed to LCM is already provider-normalized.
- */
-/**
- * Sum prompt tokens from a usage record.
- *
- * Supports two shapes:
- * - Normalized (OpenClaw internal): {input, cacheRead, cacheWrite}
- * - Raw provider: {prompt_tokens, ...}
- *
- * normalizeUsage() maps raw provider fields (prompt_tokens, cache_read, etc.)
- * to the canonical normalized shape before LCM receives runtimeContext.
- * We accept both shapes to be robust to direct test calls and future changes.
- */
-function sumPromptTokensFromUsageRecord(record: Record<string, unknown> | null): number | undefined {
-  if (!record) {
-    return undefined;
-  }
-  // Normalized shape: input + cacheRead + cacheWrite
-  const input = normalizeNonNegativeInteger(record["input"]);
-  const cacheRead = normalizeNonNegativeInteger(record["cacheRead"]);
-  const cacheWrite = normalizeNonNegativeInteger(record["cacheWrite"]);
-  if (input !== undefined || cacheRead !== undefined || cacheWrite !== undefined) {
-    return (input ?? 0) + (cacheRead ?? 0) + (cacheWrite ?? 0);
-  }
-  // Raw provider shape: prompt_tokens (already includes cache reads)
-  const rawPromptTokens = normalizeNonNegativeInteger(
-    record["prompt_tokens"] ?? record["promptTokens"] ?? record["input_tokens"] ?? record["inputTokens"],
-  );
-  if (rawPromptTokens !== undefined) {
-    return rawPromptTokens;
-  }
-  return undefined;
-}
-
-function extractRuntimePromptTokenCount(runtimeContext?: Record<string, unknown>): number | undefined {
-  const ctx = asRecord(runtimeContext);
-  if (!ctx) {
-    return undefined;
-  }
-
-  // 1. Direct currentTokenCount (already derived by OpenClaw: input+cacheRead+cacheWrite)
-  const direct = normalizeNonNegativeInteger(ctx["currentTokenCount"]);
-  if (direct !== undefined) {
-    return direct;
-  }
-
-  // 2. Sum from runtimeContext.usage (normalizeUsage output: {input, cacheRead, cacheWrite})
-  const usageSum = sumPromptTokensFromUsageRecord(
-    asRecord(ctx["usage"]) ?? asRecord(ctx["lastCallUsage"]) ?? null,
-  );
-  if (usageSum !== undefined && usageSum > 0) {
-    return usageSum;
-  }
-
-  // 3. Sum from promptCache.lastCallUsage (same normalized shape)
-  const promptCache = asRecord(ctx["promptCache"]);
-  const promptCacheUsageSum = sumPromptTokensFromUsageRecord(
-    asRecord(promptCache?.["lastCallUsage"]) ?? null,
-  );
-  if (promptCacheUsageSum !== undefined && promptCacheUsageSum > 0) {
-    return promptCacheUsageSum;
-  }
-
-  return undefined;
-}
-
-/**
- * Resolve the first-time bootstrap token budget.
- *
- * When unset, bootstrap keeps a modest suffix of the parent session rather than
- * inheriting the full raw history into a brand-new conversation.
- */
-function resolveBootstrapMaxTokens(config: Pick<LcmConfig, "bootstrapMaxTokens" | "leafChunkTokens">): number {
-  if (
-    typeof config.bootstrapMaxTokens === "number" &&
-    Number.isFinite(config.bootstrapMaxTokens) &&
-    config.bootstrapMaxTokens > 0
-  ) {
-    return Math.floor(config.bootstrapMaxTokens);
-  }
-
-  const leafChunkTokens =
-    typeof config.leafChunkTokens === "number" &&
-    Number.isFinite(config.leafChunkTokens) &&
-    config.leafChunkTokens > 0
-      ? Math.floor(config.leafChunkTokens)
-      : 40_000;
-  return Math.max(6000, Math.floor(leafChunkTokens * 0.3));
-}
-
-/**
- * Keep only the newest bootstrap messages that fit within the token budget.
- *
- * The newest message is always preserved so a fork never starts empty when the
- * parent transcript has any recoverable content at all.
- */
-function trimBootstrapMessagesToBudget(messages: AgentMessage[], maxTokens: number): AgentMessage[] {
-  if (messages.length === 0) {
-    return [];
-  }
-
-  const safeMaxTokens = Number.isFinite(maxTokens) ? Math.floor(maxTokens) : 0;
-  if (safeMaxTokens <= 0) {
-    return [messages[messages.length - 1]!];
-  }
-
-  const kept: AgentMessage[] = [];
-  let totalTokens = 0;
-
-  for (let index = messages.length - 1; index >= 0; index -= 1) {
-    const message = messages[index]!;
-    const tokenCount = toStoredMessage(message).tokenCount;
-    if (kept.length > 0 && totalTokens + tokenCount > safeMaxTokens) {
-      break;
-    }
-    kept.push(message);
-    totalTokens += tokenCount;
-  }
-
-  // If a single oversized tail message exceeds the budget, return empty
-  // rather than silently bypassing the budget cap. An empty bootstrap is
-  // safer than an exploding one.
-  if (kept.length === 1 && totalTokens > safeMaxTokens) {
-    return [];
-  }
-
-  kept.reverse();
-  return kept;
 }
 
 export type RotateSessionStorageResult =
@@ -2005,1406 +339,6 @@ type StartupAutoRotateBatchResult = {
   backupPath?: string;
   backupCreated: number;
 };
-
-function readBootstrapMessageFromJsonLine(line: string | null): AgentMessage | null {
-  if (!line) {
-    return null;
-  }
-  try {
-    return extractBootstrapMessageCandidate(JSON.parse(line));
-  } catch {
-    return null;
-  }
-}
-
-function messageIdentity(role: string, content: string): string {
-  return `${role}\u0000${content}`;
-}
-
-function isBootstrapReplayCandidateMessage(message: AgentMessage): boolean {
-  const role = toStoredMessage(message).role;
-  return role === "assistant" || role === "tool";
-}
-
-function createLosslessMessageSignature(message: AgentMessage): string {
-  const stored = toStoredMessage(message);
-  const parts = buildMessageParts({
-    sessionId: "lossless-message-signature",
-    message,
-    fallbackContent: stored.content,
-  });
-
-  return JSON.stringify({
-    role: stored.role,
-    content: stored.content,
-    parts: parts.map((part) => ({
-      partType: part.partType,
-      ordinal: part.ordinal,
-      textContent: part.textContent ?? null,
-      toolCallId: part.toolCallId ?? null,
-      toolName: part.toolName ?? null,
-      toolInput: part.toolInput ?? null,
-      toolOutput: part.toolOutput ?? null,
-      metadata: part.metadata ?? null,
-    })),
-  });
-}
-
-function createBootstrapReplaySignature(message: AgentMessage): string {
-  return createLosslessMessageSignature(message);
-}
-
-function normalizeSummaryOverlapText(value: string): string {
-  return value.replace(/\s+/g, " ").trim().toLowerCase();
-}
-
-function messageContentCoveredBySummary(params: {
-  message: AgentMessage;
-  summary: string;
-}): boolean {
-  const content = normalizeSummaryOverlapText(toStoredMessage(params.message).content);
-  if (content.length < 24) {
-    return false;
-  }
-  const summary = normalizeSummaryOverlapText(params.summary);
-  if (!summary.includes(content)) {
-    return false;
-  }
-  // Bare substring match is too loose: a 24+ char user instruction can
-  // coincidentally appear inside a long narrative summary and get silently
-  // dropped. Require one of:
-  //   1. content appears at the very start or end of the summary, OR
-  //   2. content appears inside a quoted block — double quotes ("..."),
-  //      single quotes ('...'), or backticks (`...`). All three quote
-  //      styles survive normalization and are emitted by the summarizer
-  //      when it embeds verbatim user text.
-  // Otherwise treat it as a coincidental collision and keep the message.
-  if (summary.startsWith(content) || summary.endsWith(content)) {
-    return true;
-  }
-  // Walk each quote-delimited span (cheap; summaries are bounded) and check
-  // membership. Use double-quoted literals to match the rest of the file.
-  for (const quoteChar of ["\"", "'", "`"]) {
-    let cursor = 0;
-    while (cursor < summary.length) {
-      const open = summary.indexOf(quoteChar, cursor);
-      if (open < 0) break;
-      const close = summary.indexOf(quoteChar, open + 1);
-      if (close < 0) {
-        // Unmatched opening quote: don't break out of the entire scan —
-        // a later well-formed quoted span may still contain the content.
-        // Skip past this lone opener and continue.
-        cursor = open + 1;
-        continue;
-      }
-      const span = summary.slice(open + 1, close);
-      if (span.includes(content)) {
-        return true;
-      }
-      cursor = close + 1;
-    }
-  }
-  return false;
-}
-
-const INTER_SESSION_MESSAGE_MARKER = "[Inter-session message]";
-const INTERNAL_CONTEXT_BEGIN_MARKER = "<<<BEGIN_OPENCLAW_INTERNAL_CONTEXT>>>";
-const INTERNAL_CONTEXT_END_MARKER = "<<<END_OPENCLAW_INTERNAL_CONTEXT>>>";
-const INTERNAL_TASK_COMPLETION_EVENT_MARKER = "[Internal task completion event]";
-const FALLBACK_RETRY_PROMPT_MARKER =
-  "[Retry after the previous model attempt failed or timed out]";
-const NORMALIZED_INTER_SESSION_MESSAGE_MARKER = normalizeSummaryOverlapText(INTER_SESSION_MESSAGE_MARKER);
-const NORMALIZED_INTERNAL_TASK_COMPLETION_EVENT_MARKER = normalizeSummaryOverlapText(
-  INTERNAL_TASK_COMPLETION_EVENT_MARKER,
-);
-
-function hasFallbackRetryPromptMarker(content: string): boolean {
-  return content.split(/\r?\n/).some((line) => line.trim() === FALLBACK_RETRY_PROMPT_MARKER);
-}
-
-function hasCompleteInternalContextBlock(content: string): boolean {
-  const beginIndex = content.indexOf(INTERNAL_CONTEXT_BEGIN_MARKER);
-  if (beginIndex < 0) {
-    return false;
-  }
-  return (
-    content.indexOf(
-      INTERNAL_CONTEXT_END_MARKER,
-      beginIndex + INTERNAL_CONTEXT_BEGIN_MARKER.length,
-    ) >= 0
-  );
-}
-
-function isVolatileLiveInputContent(content: string): boolean {
-  const trimmed = content.trimStart();
-  if (hasFallbackRetryPromptMarker(trimmed)) {
-    return true;
-  }
-  if (!hasCompleteInternalContextBlock(trimmed)) {
-    return false;
-  }
-  const normalized = normalizeSummaryOverlapText(trimmed);
-  if (normalized.startsWith(NORMALIZED_INTER_SESSION_MESSAGE_MARKER)) {
-    return true;
-  }
-  return (
-    trimmed.startsWith(INTERNAL_CONTEXT_BEGIN_MARKER) &&
-    normalized.includes(NORMALIZED_INTERNAL_TASK_COMPLETION_EVENT_MARKER)
-  );
-}
-
-/**
- * Estimate live message tokens for prompt-budget math.
- *
- * Serializes the full message structure (matching the model-boundary
- * renderer) instead of the stored-content token count, which omits
- * structured tool payloads and undercounts tool-heavy live messages.
- */
-function estimateAgentMessageTokens(messages: AgentMessage[]): number {
-  return estimateSerializedMessagesTokens(messages);
-}
-
-function stripTrailingAssistantPrefill(messages: AgentMessage[]): AgentMessage[] {
-  const trimmed = messages.slice();
-  while (trimmed.length > 0 && trimmed[trimmed.length - 1]?.role === "assistant") {
-    trimmed.pop();
-  }
-  return trimmed;
-}
-
-function isVolatileLiveInputMessage(message: AgentMessage): boolean {
-  const stored = toStoredMessage(message);
-  if (stored.role !== "user" && stored.role !== "system") {
-    return false;
-  }
-  if (!stored.content.trim()) {
-    return false;
-  }
-  return isVolatileLiveInputContent(stored.content);
-}
-
-function extractToolPairingIdFromRecord(record: Record<string, unknown>): string | undefined {
-  return (
-    safeString(record.toolCallId) ??
-    safeString(record.tool_call_id) ??
-    safeString(record.toolUseId) ??
-    safeString(record.tool_use_id) ??
-    safeString(record.call_id) ??
-    safeString(record.id)
-  );
-}
-
-function extractAssistantToolCallIdsForPairing(message: AgentMessage): string[] {
-  if (message.role !== "assistant" || !("content" in message) || !Array.isArray(message.content)) {
-    return [];
-  }
-  const ids: string[] = [];
-  for (const block of message.content) {
-    const record = asRecord(block);
-    if (!record || typeof record.type !== "string" || !TOOL_CALL_RAW_TYPES.has(record.type)) {
-      continue;
-    }
-    const id = extractToolPairingIdFromRecord(record);
-    if (id) {
-      ids.push(id);
-    }
-  }
-  return ids;
-}
-
-function extractToolResultIdForPairing(message: AgentMessage): string | undefined {
-  if (message.role !== "tool" && message.role !== "toolResult") {
-    return undefined;
-  }
-  const topLevel = asRecord(message);
-  if (topLevel) {
-    const direct = extractToolPairingIdFromRecord(topLevel);
-    if (direct) {
-      return direct;
-    }
-  }
-  if (!("content" in message) || !Array.isArray(message.content)) {
-    return undefined;
-  }
-  for (const block of message.content) {
-    const record = asRecord(block);
-    if (!record || typeof record.type !== "string" || !TOOL_RESULT_RAW_TYPES.has(record.type)) {
-      continue;
-    }
-    const id = extractToolPairingIdFromRecord(record);
-    if (id) {
-      return id;
-    }
-  }
-  return undefined;
-}
-
-function expandProtectedToolPairIndexes(params: {
-  assembledMessages: AgentMessage[];
-  protectedAssembledIndexes: Set<number>;
-}): Set<number> {
-  const protectedIndexes = new Set(params.protectedAssembledIndexes);
-  const assistantIndexesByToolCallId = new Map<string, number[]>();
-  const toolResultIndexesByToolCallId = new Map<string, number[]>();
-
-  for (let index = 0; index < params.assembledMessages.length; index++) {
-    const message = params.assembledMessages[index] as AgentMessage;
-    for (const toolCallId of extractAssistantToolCallIdsForPairing(message)) {
-      const indexes = assistantIndexesByToolCallId.get(toolCallId);
-      if (indexes) {
-        indexes.push(index);
-      } else {
-        assistantIndexesByToolCallId.set(toolCallId, [index]);
-      }
-    }
-    const toolResultId = extractToolResultIdForPairing(message);
-    if (toolResultId) {
-      const indexes = toolResultIndexesByToolCallId.get(toolResultId);
-      if (indexes) {
-        indexes.push(index);
-      } else {
-        toolResultIndexesByToolCallId.set(toolResultId, [index]);
-      }
-    }
-  }
-
-  let changed = true;
-  while (changed) {
-    changed = false;
-    for (let index = 0; index < params.assembledMessages.length; index++) {
-      if (!protectedIndexes.has(index)) {
-        continue;
-      }
-      const message = params.assembledMessages[index] as AgentMessage;
-      const relatedIndexes: number[] = [];
-      for (const toolCallId of extractAssistantToolCallIdsForPairing(message)) {
-        relatedIndexes.push(...(toolResultIndexesByToolCallId.get(toolCallId) ?? []));
-      }
-      const toolResultId = extractToolResultIdForPairing(message);
-      if (toolResultId) {
-        relatedIndexes.push(...(assistantIndexesByToolCallId.get(toolResultId) ?? []));
-      }
-      for (const relatedIndex of relatedIndexes) {
-        if (!protectedIndexes.has(relatedIndex)) {
-          protectedIndexes.add(relatedIndex);
-          changed = true;
-        }
-      }
-    }
-  }
-
-  return protectedIndexes;
-}
-
-function expandToolPairLiveSortIndexes(params: {
-  assembledMessages: AgentMessage[];
-  liveSortIndexes: Map<number, number>;
-}): Map<number, number> {
-  const liveSortIndexes = new Map(params.liveSortIndexes);
-  const assistantIndexesByToolCallId = new Map<string, number[]>();
-  const toolResultIndexesByToolCallId = new Map<string, number[]>();
-
-  for (let index = 0; index < params.assembledMessages.length; index++) {
-    const message = params.assembledMessages[index] as AgentMessage;
-    for (const toolCallId of extractAssistantToolCallIdsForPairing(message)) {
-      const indexes = assistantIndexesByToolCallId.get(toolCallId);
-      if (indexes) {
-        indexes.push(index);
-      } else {
-        assistantIndexesByToolCallId.set(toolCallId, [index]);
-      }
-    }
-    const toolResultId = extractToolResultIdForPairing(message);
-    if (toolResultId) {
-      const indexes = toolResultIndexesByToolCallId.get(toolResultId);
-      if (indexes) {
-        indexes.push(index);
-      } else {
-        toolResultIndexesByToolCallId.set(toolResultId, [index]);
-      }
-    }
-  }
-
-  let changed = true;
-  while (changed) {
-    changed = false;
-    for (let index = 0; index < params.assembledMessages.length; index++) {
-      const liveIndex = liveSortIndexes.get(index);
-      if (liveIndex === undefined) {
-        continue;
-      }
-      const message = params.assembledMessages[index] as AgentMessage;
-      const relatedIndexes: number[] = [];
-      for (const toolCallId of extractAssistantToolCallIdsForPairing(message)) {
-        relatedIndexes.push(...(toolResultIndexesByToolCallId.get(toolCallId) ?? []));
-      }
-      const toolResultId = extractToolResultIdForPairing(message);
-      if (toolResultId) {
-        relatedIndexes.push(...(assistantIndexesByToolCallId.get(toolResultId) ?? []));
-      }
-      for (const relatedIndex of relatedIndexes) {
-        const existing = liveSortIndexes.get(relatedIndex);
-        if (existing === undefined || liveIndex < existing) {
-          liveSortIndexes.set(relatedIndex, liveIndex);
-          changed = true;
-        }
-      }
-    }
-  }
-
-  return liveSortIndexes;
-}
-
-function buildToolPairIndexesByAssembledIndex(
-  assembledMessages: AgentMessage[],
-): Map<number, Set<number>> {
-  const assistantIndexesByToolCallId = new Map<string, number[]>();
-  const toolResultIndexesByToolCallId = new Map<string, number[]>();
-
-  // First index both sides by tool call id so matching assistant/result turns
-  // can be treated as one eviction unit.
-  for (let index = 0; index < assembledMessages.length; index++) {
-    const message = assembledMessages[index] as AgentMessage;
-    for (const toolCallId of extractAssistantToolCallIdsForPairing(message)) {
-      const indexes = assistantIndexesByToolCallId.get(toolCallId);
-      if (indexes) {
-        indexes.push(index);
-      } else {
-        assistantIndexesByToolCallId.set(toolCallId, [index]);
-      }
-    }
-    const toolResultId = extractToolResultIdForPairing(message);
-    if (toolResultId) {
-      const indexes = toolResultIndexesByToolCallId.get(toolResultId);
-      if (indexes) {
-        indexes.push(index);
-      } else {
-        toolResultIndexesByToolCallId.set(toolResultId, [index]);
-      }
-    }
-  }
-
-  const neighborsByIndex = new Map<number, Set<number>>();
-  // Link matched pairs as an undirected graph; the final groups are connected
-  // components, which handles multi-tool assistant turns and duplicate ids.
-  const linkIndexes = (left: number, right: number) => {
-    const leftNeighbors = neighborsByIndex.get(left) ?? new Set<number>([left]);
-    leftNeighbors.add(right);
-    neighborsByIndex.set(left, leftNeighbors);
-
-    const rightNeighbors = neighborsByIndex.get(right) ?? new Set<number>([right]);
-    rightNeighbors.add(left);
-    neighborsByIndex.set(right, rightNeighbors);
-  };
-
-  for (const [toolCallId, assistantIndexes] of assistantIndexesByToolCallId.entries()) {
-    const toolResultIndexes = toolResultIndexesByToolCallId.get(toolCallId) ?? [];
-    for (const assistantIndex of assistantIndexes) {
-      for (const toolResultIndex of toolResultIndexes) {
-        linkIndexes(assistantIndex, toolResultIndex);
-      }
-    }
-  }
-
-  const groupsByIndex = new Map<number, Set<number>>();
-  // Materialize every index's component so budget trimming can cheaply ask
-  // "what else must be evicted with this message?"
-  for (let index = 0; index < assembledMessages.length; index++) {
-    const group = new Set<number>();
-    const pending = [index];
-    while (pending.length > 0) {
-      const current = pending.pop() as number;
-      if (group.has(current)) {
-        continue;
-      }
-      group.add(current);
-      for (const neighbor of neighborsByIndex.get(current) ?? [current]) {
-        if (!group.has(neighbor)) {
-          pending.push(neighbor);
-        }
-      }
-    }
-    groupsByIndex.set(index, group);
-  }
-  return groupsByIndex;
-}
-
-function normalizeLiveMessageForAssemblyReconciliation(message: AgentMessage): AgentMessage {
-  const stored = toStoredMessage(message);
-  if (stored.role !== "system" && stored.role !== "tool") {
-    return message;
-  }
-  const runtimeRole = stored.role === "system" ? "user" : "toolResult";
-  const parts =
-    "content" in message
-      ? buildMessageParts({
-          sessionId: "live-reconciliation",
-          message,
-          fallbackContent: stored.content,
-        }).map((part) => toSyntheticMessagePartRecord(part, 0))
-      : [];
-  const content = contentFromParts(parts, runtimeRole, stored.content);
-  return {
-    ...message,
-    role: runtimeRole,
-    content,
-  } as AgentMessage;
-}
-
-function countNonOverlappingOccurrences(params: {
-  haystack: string;
-  needle: string;
-}): number {
-  if (!params.needle) {
-    return 0;
-  }
-  let count = 0;
-  let cursor = 0;
-  while (cursor <= params.haystack.length) {
-    const found = params.haystack.indexOf(params.needle, cursor);
-    if (found < 0) {
-      break;
-    }
-    count++;
-    cursor = found + params.needle.length;
-  }
-  return count;
-}
-
-function liveInputCoverageCapacity(params: {
-  assembledMessage: AgentMessage;
-  liveMessage: AgentMessage;
-  /**
-   * When true, the live message is a volatile input that was never persisted to
-   * DB. Summary substring coverage is insufficient for such messages because a
-   * summary that contains similar text is summarizing a *past* turn — it does
-   * not prove the current turn's volatile input is already represented.
-   */
-  isVolatileLiveInput?: boolean;
-}): number {
-  const assembled = toStoredMessage(params.assembledMessage);
-  const live = toStoredMessage(params.liveMessage);
-  if (messagesHaveSameLiveCoverageSignature(params.assembledMessage, params.liveMessage)) {
-    return 1;
-  }
-
-  // Volatile live inputs are never persisted to DB. A summary containing
-  // similar text covers a *past* occurrence, not the current live input.
-  // Only exact assembled message matches (handled above) can cover a
-  // volatile input — a historical summary paraphrase is insufficient.
-  if (params.isVolatileLiveInput) {
-    return 0;
-  }
-
-  // Substring coverage is only safe for LCM summary wrappers. Raw assembled
-  // turns must match exactly, otherwise normalized near-matches can hide a
-  // distinct volatile live input.
-  if (!assembled.content.includes("<summary ") || !assembled.content.includes("</summary>")) {
-    return 0;
-  }
-
-  const liveText = normalizeSummaryOverlapText(live.content);
-  if (liveText.length < 24) {
-    return 0;
-  }
-  const assembledText = normalizeSummaryOverlapText(assembled.content);
-  return countNonOverlappingOccurrences({ haystack: assembledText, needle: liveText });
-}
-
-function isSummaryWrapperContent(content: string): boolean {
-  return content.includes("<summary ") && content.includes("</summary>");
-}
-
-type VolatileLiveInputEntry = {
-  message: AgentMessage;
-  liveIndex: number;
-};
-
-type VolatileLiveInputCandidate = VolatileLiveInputEntry & {
-  liveText: string;
-};
-
-type VolatileLiveInputCoverageSlot = {
-  assembledIndex: number;
-};
-
-type RetainedAssembledEntry = {
-  message: AgentMessage;
-  index: number;
-};
-
-function materializeVolatileLiveInputEntries(entries: VolatileLiveInputEntry[]): AgentMessage[] {
-  return entries
-    .slice()
-    .sort((a, b) => a.liveIndex - b.liveIndex)
-    .map((entry) => entry.message);
-}
-
-function hashAgentMessageForAssemblyProtection(message: AgentMessage): string {
-  return createHash("sha256").update(JSON.stringify([message])).digest("hex").slice(0, 16);
-}
-
-function resolveProtectedFreshTailAssembledIndexes(params: {
-  assembledMessages: AgentMessage[];
-  freshTailMessageHashes?: string[];
-}): Set<number> {
-  const protectedIndexes = new Set<number>();
-  const usedIndexes = new Set<number>();
-  for (const hash of params.freshTailMessageHashes ?? []) {
-    for (let index = params.assembledMessages.length - 1; index >= 0; index--) {
-      if (usedIndexes.has(index)) {
-        continue;
-      }
-      const message = params.assembledMessages[index] as AgentMessage;
-      if (hashAgentMessageForAssemblyProtection(message) === hash) {
-        protectedIndexes.add(index);
-        usedIndexes.add(index);
-        break;
-      }
-    }
-  }
-  return protectedIndexes;
-}
-
-function messagesHaveSameLosslessSignature(left: AgentMessage, right: AgentMessage): boolean {
-  return createLosslessMessageSignature(left) === createLosslessMessageSignature(right);
-}
-
-function createLiveCoverageSignature(message: AgentMessage): string {
-  const stored = toStoredMessage(message);
-  if (
-    (stored.role === "user" || stored.role === "system" || stored.role === "assistant") &&
-    stored.content.length > 0 &&
-    isCanonicalTextOnlyMessage(message, stored.content)
-  ) {
-    return JSON.stringify({
-      kind: "canonical-text",
-      role: stored.role,
-      content: stored.content,
-    });
-  }
-  const canonicalToolTextSignature = createCanonicalToolTextCoverageSignature(
-    message,
-    stored.content,
-  );
-  if (canonicalToolTextSignature) {
-    return canonicalToolTextSignature;
-  }
-  return createLosslessMessageSignature(message);
-}
-
-function normalizeToolNameForCoverage(toolName: string | null | undefined): string | null {
-  // The assembler fills missing tool names with "unknown" on rehydration.
-  // Treat null/undefined/""/"unknown" as equivalent for coverage matching
-  // so live and assembled tool-result signatures still match.
-  if (!toolName || toolName === "unknown") {
-    return null;
-  }
-  return toolName;
-}
-
-function createCanonicalToolTextCoverageSignature(
-  message: AgentMessage,
-  fallbackContent: string,
-): string | undefined {
-  const stored = toStoredMessage(message);
-  if (stored.role !== "tool" || fallbackContent.length === 0) {
-    return undefined;
-  }
-  const parts = buildMessageParts({
-    sessionId: "live-tool-coverage-signature",
-    message,
-    fallbackContent,
-  });
-  if (parts.length !== 1) {
-    return undefined;
-  }
-  const part = parts[0] as CreateMessagePartInput;
-  if (
-    part.partType !== "text" ||
-    (part.textContent ?? "") !== fallbackContent ||
-    part.toolInput != null ||
-    part.toolOutput != null
-  ) {
-    return undefined;
-  }
-  return JSON.stringify({
-    kind: "canonical-tool-text",
-    role: stored.role,
-    content: fallbackContent,
-    toolCallId: part.toolCallId ?? extractToolResultIdForPairing(message) ?? null,
-    toolName: normalizeToolNameForCoverage(part.toolName),
-  });
-}
-
-function isCanonicalTextOnlyMessage(message: AgentMessage, fallbackContent: string): boolean {
-  const parts = buildMessageParts({
-    sessionId: "live-coverage-signature",
-    message,
-    fallbackContent,
-  });
-  if (parts.length !== 1) {
-    return false;
-  }
-  const part = parts[0] as CreateMessagePartInput;
-  return (
-    part.partType === "text" &&
-    (part.textContent ?? "") === fallbackContent &&
-    part.toolCallId == null &&
-    part.toolName == null &&
-    part.toolInput == null &&
-    part.toolOutput == null
-  );
-}
-
-function messagesHaveSameLiveCoverageSignature(left: AgentMessage, right: AgentMessage): boolean {
-  return createLiveCoverageSignature(left) === createLiveCoverageSignature(right);
-}
-
-function resolveExactAssembledLiveSortIndexes(params: {
-  assembledMessages: AgentMessage[];
-  liveMessages: AgentMessage[];
-}): Map<number, number> {
-  const liveSortIndexes = new Map<number, number>();
-  const usedAssembledIndexes = new Set<number>();
-  for (let liveIndex = params.liveMessages.length - 1; liveIndex >= 0; liveIndex--) {
-    const liveMessage = params.liveMessages[liveIndex] as AgentMessage;
-    for (
-      let assembledIndex = params.assembledMessages.length - 1;
-      assembledIndex >= 0;
-      assembledIndex--
-    ) {
-      if (usedAssembledIndexes.has(assembledIndex)) {
-        continue;
-      }
-      const assembledMessage = params.assembledMessages[assembledIndex] as AgentMessage;
-      if (messagesHaveSameLiveCoverageSignature(assembledMessage, liveMessage)) {
-        liveSortIndexes.set(assembledIndex, liveIndex);
-        usedAssembledIndexes.add(assembledIndex);
-        break;
-      }
-    }
-  }
-  return liveSortIndexes;
-}
-
-function mergeCoveredVolatileLiveSortIndexes(params: {
-  exactLiveSortIndexes: Map<number, number>;
-  coveredEntriesByAssembledIndex: Map<number, VolatileLiveInputEntry[]>;
-}): Map<number, number> {
-  const liveSortIndexes = new Map(params.exactLiveSortIndexes);
-  for (const [assembledIndex, entries] of params.coveredEntriesByAssembledIndex.entries()) {
-    const coveredLiveIndex = Math.min(...entries.map((entry) => entry.liveIndex));
-    const existingLiveIndex = liveSortIndexes.get(assembledIndex);
-    if (existingLiveIndex === undefined || coveredLiveIndex < existingLiveIndex) {
-      liveSortIndexes.set(assembledIndex, coveredLiveIndex);
-    }
-  }
-  return liveSortIndexes;
-}
-
-function buildVolatileLiveInputMergedOutput(params: {
-  retained: RetainedAssembledEntry[];
-  appendedEntries: VolatileLiveInputEntry[];
-  liveSortIndexes: Map<number, number>;
-  log?: RepairLogger;
-}): AgentMessage[] {
-  const output: AgentMessage[] = [];
-  const appendedEntries = params.appendedEntries
-    .slice()
-    .sort((left, right) => left.liveIndex - right.liveIndex);
-  let appendedCursor = 0;
-  for (const retainedEntry of params.retained) {
-    const retainedLiveIndex = params.liveSortIndexes.get(retainedEntry.index);
-    if (retainedLiveIndex !== undefined) {
-      while (
-        appendedCursor < appendedEntries.length &&
-        (appendedEntries[appendedCursor] as VolatileLiveInputEntry).liveIndex < retainedLiveIndex
-      ) {
-        output.push((appendedEntries[appendedCursor] as VolatileLiveInputEntry).message);
-        appendedCursor++;
-      }
-    }
-    output.push(retainedEntry.message);
-  }
-  while (appendedCursor < appendedEntries.length) {
-    output.push((appendedEntries[appendedCursor] as VolatileLiveInputEntry).message);
-    appendedCursor++;
-  }
-  return sanitizeToolUseResultPairing(output, params.log) as AgentMessage[];
-}
-
-function matchVolatileLiveInputsToCoverageSlots(params: {
-  assembledMessages: AgentMessage[];
-  volatileLiveInputs: VolatileLiveInputCandidate[];
-}): Map<number, number> {
-  const entryIndexesByLiveText = new Map<string, number[]>();
-  for (let entryIndex = 0; entryIndex < params.volatileLiveInputs.length; entryIndex++) {
-    const entry = params.volatileLiveInputs[entryIndex] as VolatileLiveInputCandidate;
-    const entryIndexes = entryIndexesByLiveText.get(entry.liveText);
-    if (entryIndexes) {
-      entryIndexes.push(entryIndex);
-    } else {
-      entryIndexesByLiveText.set(entry.liveText, [entryIndex]);
-    }
-  }
-
-  const slots: VolatileLiveInputCoverageSlot[] = [];
-  const candidateSlotIndexesByEntryIndex = params.volatileLiveInputs.map(() => [] as number[]);
-  const addCandidateSlots = (entryIndexes: number[], assembledIndex: number, slotCount: number) => {
-    const slotIndexes: number[] = [];
-    for (let slotOffset = 0; slotOffset < slotCount; slotOffset++) {
-      slotIndexes.push(slots.length);
-      slots.push({ assembledIndex });
-    }
-    for (const entryIndex of entryIndexes) {
-      candidateSlotIndexesByEntryIndex[entryIndex]?.push(...slotIndexes);
-    }
-  };
-
-  for (const [liveText, entryIndexes] of entryIndexesByLiveText.entries()) {
-    for (let assembledIndex = 0; assembledIndex < params.assembledMessages.length; assembledIndex++) {
-      const assembledMessage = params.assembledMessages[assembledIndex] as AgentMessage;
-      const assembled = toStoredMessage(assembledMessage);
-      if (!isSummaryWrapperContent(assembled.content)) {
-        const exactEntryIndexes = entryIndexes.filter((entryIndex) =>
-          messagesHaveSameLiveCoverageSignature(
-            assembledMessage,
-            (params.volatileLiveInputs[entryIndex] as VolatileLiveInputCandidate).message,
-          )
-        );
-        if (exactEntryIndexes.length > 0) {
-          addCandidateSlots(exactEntryIndexes, assembledIndex, 1);
-        }
-        continue;
-      }
-
-      const representativeEntry = params.volatileLiveInputs[entryIndexes[0] as number] as VolatileLiveInputCandidate;
-      const capacity = liveInputCoverageCapacity({
-        assembledMessage,
-        liveMessage: representativeEntry.message,
-        isVolatileLiveInput: true,
-      });
-      if (capacity <= 0) {
-        continue;
-      }
-
-      const entryIndexesBySignature = new Map<string, number[]>();
-      for (const entryIndex of entryIndexes) {
-        const entry = params.volatileLiveInputs[entryIndex] as VolatileLiveInputCandidate;
-        const signature = createLiveCoverageSignature(entry.message);
-        const signatureEntryIndexes = entryIndexesBySignature.get(signature);
-        if (signatureEntryIndexes) {
-          signatureEntryIndexes.push(entryIndex);
-        } else {
-          entryIndexesBySignature.set(signature, [entryIndex]);
-        }
-      }
-
-      let exactSlotCount = 0;
-      for (const signatureEntryIndexes of entryIndexesBySignature.values()) {
-        const firstEntry = params.volatileLiveInputs[signatureEntryIndexes[0] as number] as VolatileLiveInputCandidate;
-        const liveContent = toStoredMessage(firstEntry.message).content;
-        const exactCapacity = liveContent
-          ? countNonOverlappingOccurrences({ haystack: assembled.content, needle: liveContent })
-          : 0;
-        const slotCount = Math.min(exactCapacity, signatureEntryIndexes.length);
-        if (slotCount > 0) {
-          addCandidateSlots(signatureEntryIndexes, assembledIndex, slotCount);
-          exactSlotCount += slotCount;
-        }
-      }
-
-      const genericSlotCount = Math.min(
-        Math.max(0, capacity - exactSlotCount),
-        Math.max(0, entryIndexes.length - exactSlotCount),
-      );
-      if (genericSlotCount > 0) {
-        addCandidateSlots(entryIndexes, assembledIndex, genericSlotCount);
-      }
-    }
-  }
-
-  const slotToEntryIndex = new Map<number, number>();
-  const tryAssignEntry = (entryIndex: number, visitedSlots: Set<number>): boolean => {
-    const candidateSlotIndexes = candidateSlotIndexesByEntryIndex[entryIndex] ?? [];
-    for (const slotIndex of candidateSlotIndexes) {
-      if (visitedSlots.has(slotIndex)) {
-        continue;
-      }
-      visitedSlots.add(slotIndex);
-      const currentEntryIndex = slotToEntryIndex.get(slotIndex);
-      if (
-        currentEntryIndex === undefined ||
-        tryAssignEntry(currentEntryIndex, visitedSlots)
-      ) {
-        slotToEntryIndex.set(slotIndex, entryIndex);
-        return true;
-      }
-    }
-    return false;
-  };
-
-  for (let entryIndex = 0; entryIndex < params.volatileLiveInputs.length; entryIndex++) {
-    tryAssignEntry(entryIndex, new Set<number>());
-  }
-
-  const entryToAssembledIndex = new Map<number, number>();
-  for (const [slotIndex, entryIndex] of slotToEntryIndex.entries()) {
-    const slot = slots[slotIndex] as VolatileLiveInputCoverageSlot;
-    entryToAssembledIndex.set(entryIndex, slot.assembledIndex);
-  }
-  return entryToAssembledIndex;
-}
-
-function collectUncoveredVolatileLiveInputs(params: {
-  assembledMessages: AgentMessage[];
-  liveMessages: AgentMessage[];
-}): {
-  entries: VolatileLiveInputEntry[];
-  estimatedTokens: number;
-  coveredEntriesByAssembledIndex: Map<number, VolatileLiveInputEntry[]>;
-} {
-  const volatileLiveInputs = params.liveMessages
-    .map((message, liveIndex) => ({ message, liveIndex }))
-    .filter((entry) => isVolatileLiveInputMessage(entry.message))
-    .map((entry) => ({
-      ...entry,
-      liveText: normalizeSummaryOverlapText(toStoredMessage(entry.message).content),
-    }));
-  const uncovered: VolatileLiveInputEntry[] = [];
-  const coveredEntriesByAssembledIndex = new Map<number, VolatileLiveInputEntry[]>();
-  const entryToAssembledIndex = matchVolatileLiveInputsToCoverageSlots({
-    assembledMessages: params.assembledMessages,
-    volatileLiveInputs,
-  });
-
-  for (let entryIndex = 0; entryIndex < volatileLiveInputs.length; entryIndex++) {
-    const entry = volatileLiveInputs[entryIndex] as VolatileLiveInputCandidate;
-    const assembledIndex = entryToAssembledIndex.get(entryIndex);
-    if (assembledIndex !== undefined) {
-      const coveredEntries = coveredEntriesByAssembledIndex.get(assembledIndex);
-      if (coveredEntries) {
-        coveredEntries.push(entry);
-      } else {
-        coveredEntriesByAssembledIndex.set(assembledIndex, [entry]);
-      }
-    } else {
-      uncovered.push(entry);
-    }
-  }
-
-  return {
-    entries: uncovered,
-    estimatedTokens: estimateAgentMessageTokens(materializeVolatileLiveInputEntries(uncovered)),
-    coveredEntriesByAssembledIndex,
-  };
-}
-
-function appendUncoveredVolatileLiveInputsWithinBudget(params: {
-  assembledMessages: AgentMessage[];
-  assembledEstimatedTokens: number;
-  liveMessages: AgentMessage[];
-  protectedAssembledIndexes?: Set<number>;
-  tokenBudget: number;
-  log?: RepairLogger;
-}): {
-  messages: AgentMessage[];
-  estimatedTokens: number;
-  appendedMessages: number;
-  appendedTokens: number;
-  evictedMessages: number;
-  evictedTokens: number;
-  overBudget: boolean;
-} {
-  const liveMessages = params.liveMessages.map(normalizeLiveMessageForAssemblyReconciliation);
-  const protectedAssembledIndexes = expandProtectedToolPairIndexes({
-    assembledMessages: params.assembledMessages,
-    protectedAssembledIndexes: params.protectedAssembledIndexes ?? new Set<number>(),
-  });
-  const uncovered = collectUncoveredVolatileLiveInputs({
-    assembledMessages: params.assembledMessages,
-    liveMessages,
-  });
-  if (uncovered.entries.length === 0) {
-    return {
-      messages: params.assembledMessages,
-      estimatedTokens: params.assembledEstimatedTokens,
-      appendedMessages: 0,
-      appendedTokens: 0,
-      evictedMessages: 0,
-      evictedTokens: 0,
-      overBudget: params.assembledEstimatedTokens > params.tokenBudget,
-    };
-  }
-
-  let retained = params.assembledMessages.map((message, index) => ({ message, index }));
-  let appendedEntries = uncovered.entries.slice();
-  const toolPairIndexesByIndex = buildToolPairIndexesByAssembledIndex(params.assembledMessages);
-  const exactLiveSortIndexes = resolveExactAssembledLiveSortIndexes({
-    assembledMessages: params.assembledMessages,
-    liveMessages,
-  });
-  const exactLiveProtectedIndexes = expandProtectedToolPairIndexes({
-    assembledMessages: params.assembledMessages,
-    protectedAssembledIndexes: new Set(exactLiveSortIndexes.keys()),
-  });
-  const liveSortIndexes = expandToolPairLiveSortIndexes({
-    assembledMessages: params.assembledMessages,
-    liveSortIndexes: mergeCoveredVolatileLiveSortIndexes({
-      exactLiveSortIndexes,
-      coveredEntriesByAssembledIndex: uncovered.coveredEntriesByAssembledIndex,
-    }),
-  });
-  let evictedMessages = 0;
-  let evictedTokens = 0;
-  let output = buildVolatileLiveInputMergedOutput({
-    retained,
-    appendedEntries,
-    liveSortIndexes,
-  });
-  let estimatedTokens = estimateAgentMessageTokens(output);
-
-  while (retained.length > 0 && estimatedTokens > params.tokenBudget) {
-    let bestCandidate:
-      | {
-          evictAssembledIndexes: Set<number>;
-          output: AgentMessage[];
-          estimatedTokens: number;
-          appendedEntries: VolatileLiveInputEntry[];
-        }
-      | undefined;
-    for (let evictIndex = 0; evictIndex < retained.length; evictIndex++) {
-      const entry = retained[evictIndex] as RetainedAssembledEntry;
-      const evictAssembledIndexes = toolPairIndexesByIndex.get(entry.index) ?? new Set([entry.index]);
-      const candidateEvictsExactLiveTurn = Array.from(evictAssembledIndexes).some((index) =>
-        exactLiveProtectedIndexes.has(index)
-      );
-      const candidateEvictsProtectedTurn = Array.from(evictAssembledIndexes).some((index) =>
-        protectedAssembledIndexes.has(index)
-      );
-      if (candidateEvictsExactLiveTurn || candidateEvictsProtectedTurn) {
-        continue;
-      }
-      const restoredCoveredEntries = Array.from(evictAssembledIndexes).flatMap(
-        (index) => uncovered.coveredEntriesByAssembledIndex.get(index) ?? [],
-      );
-      const candidateRetained = retained.filter(
-        (retainedEntry) => !evictAssembledIndexes.has(retainedEntry.index),
-      );
-      const candidateAppendedEntries =
-        restoredCoveredEntries.length > 0
-          ? [...appendedEntries, ...restoredCoveredEntries]
-          : appendedEntries;
-      const candidateOutput = buildVolatileLiveInputMergedOutput({
-        retained: candidateRetained,
-        appendedEntries: candidateAppendedEntries,
-        liveSortIndexes,
-      });
-      const candidateEstimatedTokens = estimateAgentMessageTokens(candidateOutput);
-      const candidateFits = candidateEstimatedTokens <= params.tokenBudget;
-      const bestFits =
-        bestCandidate !== undefined && bestCandidate.estimatedTokens <= params.tokenBudget;
-      if (
-        bestCandidate === undefined ||
-        (candidateFits && !bestFits) ||
-        (candidateFits &&
-          bestFits &&
-          candidateEstimatedTokens > bestCandidate.estimatedTokens) ||
-        (!candidateFits &&
-          !bestFits &&
-          candidateEstimatedTokens < bestCandidate.estimatedTokens)
-      ) {
-        bestCandidate = {
-          evictAssembledIndexes,
-          output: candidateOutput,
-          estimatedTokens: candidateEstimatedTokens,
-          appendedEntries: candidateAppendedEntries,
-        };
-      }
-    }
-    if (!bestCandidate) {
-      break;
-    }
-    const removedEntries = retained.filter((entry) =>
-      bestCandidate.evictAssembledIndexes.has(entry.index),
-    );
-    retained = retained.filter((entry) => !bestCandidate.evictAssembledIndexes.has(entry.index));
-    appendedEntries = bestCandidate.appendedEntries;
-    for (const removed of removedEntries) {
-      uncovered.coveredEntriesByAssembledIndex.delete(removed.index);
-      evictedTokens += toStoredMessage(removed.message).tokenCount;
-    }
-    evictedMessages += removedEntries.length;
-    output = bestCandidate.output;
-    estimatedTokens = bestCandidate.estimatedTokens;
-  }
-  output = buildVolatileLiveInputMergedOutput({
-    retained,
-    appendedEntries,
-    liveSortIndexes,
-    log: params.log,
-  });
-  estimatedTokens = estimateAgentMessageTokens(output);
-  const appendedMessages = materializeVolatileLiveInputEntries(appendedEntries);
-
-  return {
-    messages: output,
-    estimatedTokens,
-    appendedMessages: appendedMessages.length,
-    appendedTokens: estimateAgentMessageTokens(appendedMessages),
-    evictedMessages,
-    evictedTokens,
-    overBudget: estimatedTokens > params.tokenBudget,
-  };
-}
-
-function resolveForkBoundedLiveSuffix(params: {
-  assembledMessages: AgentMessage[];
-  liveMessages: AgentMessage[];
-  forkSourceMessageCount: number;
-}): AgentMessage[] {
-  const liveMessages = params.liveMessages.map(normalizeLiveMessageForAssemblyReconciliation);
-  const forkSourceMessageCount = Math.max(0, Math.floor(params.forkSourceMessageCount));
-  const anchorSearchEnd =
-    forkSourceMessageCount > 0
-      ? Math.min(liveMessages.length, forkSourceMessageCount)
-      : liveMessages.length;
-  let anchorLiveIndex = -1;
-  for (let liveIndex = anchorSearchEnd - 1; liveIndex >= 0; liveIndex--) {
-    const liveMessage = liveMessages[liveIndex] as AgentMessage;
-    for (
-      let assembledIndex = params.assembledMessages.length - 1;
-      assembledIndex >= 0;
-      assembledIndex--
-    ) {
-      const assembledMessage = params.assembledMessages[assembledIndex] as AgentMessage;
-      if (messagesHaveSameLiveCoverageSignature(assembledMessage, liveMessage)) {
-        anchorLiveIndex = liveIndex;
-        break;
-      }
-    }
-    if (anchorLiveIndex >= 0) {
-      break;
-    }
-  }
-
-  if (anchorLiveIndex >= 0) {
-    return liveMessages.slice(anchorLiveIndex + 1);
-  }
-
-  if (forkSourceMessageCount > 0 && liveMessages.length >= forkSourceMessageCount) {
-    return liveMessages.slice(forkSourceMessageCount);
-  }
-
-  // If the host provides a short live snapshot rather than the copied fork
-  // branch, keep that snapshot; it is no longer the raw parent prefix.
-  if (forkSourceMessageCount > 0 && liveMessages.length < forkSourceMessageCount) {
-    return liveMessages;
-  }
-
-  return [];
-}
-
-/**
- * Suffix-trim live messages for prompt bounding, measured by serialized
- * (model-boundary) token estimate.
- *
- * Unlike `trimBootstrapMessagesToBudget` (which selects what to *persist*
- * during bootstrap and stays on stored-content counts), this bounds what is
- * *sent to the model*, so it must count structured tool payloads that the
- * stored-content estimate omits.
- */
-function trimMessagesToBudget(messages: AgentMessage[], tokenBudget: number): AgentMessage[] {
-  const safeMaxTokens = Number.isFinite(tokenBudget) ? Math.max(0, Math.floor(tokenBudget)) : 0;
-  if (messages.length === 0) {
-    return [];
-  }
-  if (safeMaxTokens <= 0) {
-    return stripTrailingAssistantPrefill([messages[messages.length - 1]!]);
-  }
-  const kept: AgentMessage[] = [];
-  let totalTokens = 0;
-  for (let index = messages.length - 1; index >= 0; index -= 1) {
-    const message = messages[index]!;
-    const tokenCount = estimateSerializedMessageTokens(message);
-    if (kept.length > 0 && totalTokens + tokenCount > safeMaxTokens) {
-      break;
-    }
-    kept.push(message);
-    totalTokens += tokenCount;
-  }
-  // A single oversized tail message exceeding the budget returns empty,
-  // matching the bootstrap trim contract callers already handle.
-  if (kept.length === 1 && totalTokens > safeMaxTokens) {
-    return [];
-  }
-  kept.reverse();
-  return stripTrailingAssistantPrefill(kept);
-}
-
-/**
- * Safety ratio applied to the assembly token budget when clamping final
- * output by serialized (model-boundary) estimate. Leaves headroom for the
- * host's reserve tokens and renderer overhead beyond our approximation.
- */
-const SERIALIZED_OUTPUT_CLAMP_SAFETY_RATIO = 0.9;
-
-/**
- * Final budget clamp on assembled output, measured by serialized
- * (model-boundary) token estimate rather than stored-content counts.
- *
- * Assembly budgets enforced on stored token counts can diverge from the
- * real prompt when live message objects carry structured payloads that
- * stored content omits (e.g. transcripts imported from a previous harness).
- * This clamp keeps the newest suffix that fits, drops leading tool results
- * orphaned by eviction, and re-seats the most recent user turn if eviction
- * removed every user message.
- */
-function clampMessagesToSerializedBudget(params: {
-  messages: AgentMessage[];
-  tokenBudget: number;
-}): {
-  messages: AgentMessage[];
-  serializedTokens: number;
-  serializedTokensBefore: number;
-  clamped: boolean;
-  evictedMessages: number;
-  overBudget: boolean;
-} {
-  // Trigger only when the serialized estimate actually exceeds the budget;
-  // once eviction is unavoidable, clamp down to the safety target so the
-  // result leaves headroom for host reserve tokens and renderer overhead.
-  const triggerTokens = Math.max(1, Math.floor(params.tokenBudget));
-  const targetTokens = Math.max(
-    1,
-    Math.floor(params.tokenBudget * SERIALIZED_OUTPUT_CLAMP_SAFETY_RATIO),
-  );
-  const serializedTokensBefore = estimateSerializedMessagesTokens(params.messages);
-  if (serializedTokensBefore <= triggerTokens || params.messages.length === 0) {
-    return {
-      messages: params.messages,
-      serializedTokens: serializedTokensBefore,
-      serializedTokensBefore,
-      clamped: false,
-      evictedMessages: 0,
-      overBudget: serializedTokensBefore > triggerTokens,
-    };
-  }
-
-  // Keep the newest suffix that fits the target (always at least one message).
-  const kept: AgentMessage[] = [];
-  let keptTokens = 0;
-  for (let index = params.messages.length - 1; index >= 0; index -= 1) {
-    const message = params.messages[index]!;
-    const tokenCount = estimateSerializedMessageTokens(message);
-    if (kept.length > 0 && keptTokens + tokenCount > targetTokens) {
-      break;
-    }
-    kept.push(message);
-    keptTokens += tokenCount;
-  }
-  kept.reverse();
-
-  // Eviction may have removed the assistant tool_use partner of leading
-  // tool results; drop those orphans rather than ship unpaired results.
-  while (kept.length > 1 && toRuntimeRoleForTokenEstimate(kept[0]!.role) === "toolResult") {
-    keptTokens -= estimateSerializedMessageTokens(kept[0]!);
-    kept.shift();
-  }
-
-  // The provider rejects contexts with no user turn; re-seat the most
-  // recent evicted user message if the suffix lost every one of them.
-  if (!kept.some((message) => toRuntimeRoleForTokenEstimate(message.role) === "user")) {
-    for (let index = params.messages.length - kept.length - 1; index >= 0; index -= 1) {
-      const candidate = params.messages[index]!;
-      if (toRuntimeRoleForTokenEstimate(candidate.role) === "user") {
-        kept.unshift(candidate);
-        keptTokens += estimateSerializedMessageTokens(candidate);
-        break;
-      }
-    }
-  }
-
-  // If stripping the assistant tail would empty the result (a transcript
-  // with no user turns at all), keep the unstripped suffix instead; the
-  // estimate must describe whichever set is actually returned.
-  const stripped = stripTrailingAssistantPrefill(kept);
-  const finalMessages = stripped.length > 0 ? stripped : kept;
-  const serializedTokens =
-    finalMessages.length === kept.length
-      ? keptTokens
-      : estimateSerializedMessagesTokens(finalMessages);
-  return {
-    messages: finalMessages,
-    serializedTokens,
-    serializedTokensBefore,
-    clamped: true,
-    evictedMessages: params.messages.length - finalMessages.length,
-    overBudget: serializedTokens > targetTokens,
-  };
-}
-
-function isProtectedLeadingLiveContextMessage(message: AgentMessage): boolean {
-  const role = typeof message.role === "string" ? message.role.toLowerCase() : "";
-  return role === "system" || role === "developer";
-}
-
-function buildDegradedLiveAssembleResult(params: {
-  liveMessages: AgentMessage[];
-  tokenBudget: number;
-}): AssembleResult {
-  const withoutAssistantPrefill = stripTrailingAssistantPrefill(params.liveMessages.slice());
-  const protectedPrefix: AgentMessage[] = [];
-  while (
-    protectedPrefix.length < withoutAssistantPrefill.length &&
-    isProtectedLeadingLiveContextMessage(withoutAssistantPrefill[protectedPrefix.length]!)
-  ) {
-    protectedPrefix.push(withoutAssistantPrefill[protectedPrefix.length]!);
-  }
-  const liveTail = withoutAssistantPrefill.slice(protectedPrefix.length);
-  const remainingBudget = Math.max(
-    0,
-    Math.floor(params.tokenBudget) - estimateAgentMessageTokens(protectedPrefix),
-  );
-  let liveTailMessages = trimMessagesToBudget(liveTail, remainingBudget);
-  if (liveTailMessages.length === 0 && liveTail.length > 0) {
-    liveTailMessages = [liveTail[liveTail.length - 1]!];
-  }
-  const messages = [...protectedPrefix, ...liveTailMessages];
-  return {
-    messages,
-    estimatedTokens: estimateAgentMessageTokens(messages),
-  };
-}
-
-function resolveDeferredAssemblyPressure(params: {
-  liveContextTokens: number;
-  maintenance: ConversationCompactionMaintenanceRecord | null;
-}): {
-  observedContextTokens: number;
-  projectedTokenCount: number | null;
-  pressureTokenCount: number;
-} {
-  const recordedContextTokens = normalizeNonNegativeInteger(
-    params.maintenance?.currentTokenCount,
-  );
-  const recordedProjectedTokens = normalizeNonNegativeInteger(
-    params.maintenance?.projectedTokenCount,
-  );
-  const observedContextTokens = Math.max(
-    params.liveContextTokens,
-    recordedContextTokens ?? 0,
-  );
-  const pressureTokenCount = Math.max(
-    observedContextTokens,
-    recordedProjectedTokens ?? 0,
-  );
-  return {
-    observedContextTokens,
-    projectedTokenCount: recordedProjectedTokens ?? null,
-    pressureTokenCount,
-  };
-}
-
-function buildForkBoundedLiveFallback(params: {
-  liveMessages: AgentMessage[];
-  forkSourceMessageCount: number;
-  tokenBudget: number;
-  bootstrapMaxTokens: number;
-}): AssembleResult {
-  const suffix = resolveForkBoundedLiveSuffix({
-    assembledMessages: [],
-    liveMessages: params.liveMessages,
-    forkSourceMessageCount: params.forkSourceMessageCount,
-  });
-  const candidateMessages = suffix.length > 0 ? suffix : params.liveMessages;
-  const boundedMessages = trimMessagesToBudget(
-    candidateMessages,
-    Math.min(params.tokenBudget, params.bootstrapMaxTokens),
-  );
-  return {
-    messages: boundedMessages,
-    estimatedTokens: estimateAgentMessageTokens(boundedMessages),
-  };
-}
-
-function appendForkBoundedLiveSuffixWithinBudget(params: {
-  assembledMessages: AgentMessage[];
-  assembledEstimatedTokens: number;
-  liveMessages: AgentMessage[];
-  forkSourceMessageCount: number;
-  tokenBudget: number;
-}): {
-  messages: AgentMessage[];
-  estimatedTokens: number;
-  appendedMessages: number;
-  appendedTokens: number;
-  evictedMessages: number;
-  evictedTokens: number;
-  overBudget: boolean;
-  protectedIndexes: Set<number>;
-} {
-  const suffix = stripTrailingAssistantPrefill(
-    resolveForkBoundedLiveSuffix({
-      assembledMessages: params.assembledMessages,
-      liveMessages: params.liveMessages,
-      forkSourceMessageCount: params.forkSourceMessageCount,
-    }),
-  );
-  if (suffix.length === 0) {
-    return {
-      messages: params.assembledMessages,
-      estimatedTokens: params.assembledEstimatedTokens,
-      appendedMessages: 0,
-      appendedTokens: 0,
-      evictedMessages: 0,
-      evictedTokens: 0,
-      overBudget: params.assembledEstimatedTokens > params.tokenBudget,
-      protectedIndexes: new Set(),
-    };
-  }
-
-  let retained = params.assembledMessages.slice();
-  let retainedSuffix = suffix.slice();
-  let evictedMessages = 0;
-  let evictedTokens = 0;
-  let output = [...retained, ...retainedSuffix];
-  let estimatedTokens = estimateAgentMessageTokens(output);
-
-  while (retained.length > 0 && estimatedTokens > params.tokenBudget) {
-    const removed = retained.shift() as AgentMessage;
-    evictedMessages += 1;
-    evictedTokens += toStoredMessage(removed).tokenCount;
-    output = [...retained, ...retainedSuffix];
-    estimatedTokens = estimateAgentMessageTokens(output);
-  }
-
-  while (retainedSuffix.length > 0 && estimatedTokens > params.tokenBudget) {
-    const removed = retainedSuffix.shift() as AgentMessage;
-    evictedMessages += 1;
-    evictedTokens += toStoredMessage(removed).tokenCount;
-    output = [...retained, ...retainedSuffix];
-    estimatedTokens = estimateAgentMessageTokens(output);
-  }
-
-  const protectedIndexes = new Set<number>();
-  const suffixStartIndex = output.length - retainedSuffix.length;
-  for (let index = suffixStartIndex; index < output.length; index += 1) {
-    protectedIndexes.add(index);
-  }
-
-  return {
-    messages: output,
-    estimatedTokens,
-    appendedMessages: retainedSuffix.length,
-    appendedTokens: estimateAgentMessageTokens(retainedSuffix),
-    evictedMessages,
-    evictedTokens,
-    overBudget: estimatedTokens > params.tokenBudget,
-    protectedIndexes,
-  };
-}
 
 // ── LcmContextEngine ────────────────────────────────────────────────────────
 
@@ -12744,117 +9678,6 @@ export class LcmContextEngine implements ContextEngine {
 }
 
 // ── Heartbeat detection ─────────────────────────────────────────────────────
-
-const HEARTBEAT_OK_TOKEN = "heartbeat_ok";
-const HEARTBEAT_TURN_MARKER = "heartbeat.md";
-const OPENCLAW_HEARTBEAT_POLL = "[openclaw heartbeat poll]";
-
-/**
- * Detect whether an assistant message is a heartbeat ack.
- *
- * Only exact (case-insensitive) "HEARTBEAT_OK" acknowledgements are pruned.
- * Any additional text indicates the heartbeat carried real content and should remain.
- */
-function isHeartbeatOkContent(content: string): boolean {
-  return content.trim().toLowerCase() === HEARTBEAT_OK_TOKEN;
-}
-
-/**
- * Synthetic heartbeat traffic (poll prompts and HEARTBEAT_OK acks) recurs
- * identically in every session, so it can never discriminate lineage in
- * the ambiguous-rollover freshness check.
- */
-function isHeartbeatNoiseContent(role: string, content: string): boolean {
-  const normalized = content.trim().toLowerCase();
-  if (role === "user" && normalized === OPENCLAW_HEARTBEAT_POLL) {
-    return true;
-  }
-  return role === "assistant" && normalized === HEARTBEAT_OK_TOKEN;
-}
-
-function batchLooksLikeHeartbeatAckTurn(messages: AgentMessage[]): boolean {
-  let sawHeartbeatMarker = false;
-  let sawHeartbeatAck = false;
-
-  for (const message of messages) {
-    const stored = toStoredMessage(message);
-    if (!sawHeartbeatMarker && stored.content.toLowerCase().includes(HEARTBEAT_TURN_MARKER)) {
-      sawHeartbeatMarker = true;
-    }
-    if (!sawHeartbeatAck && stored.role === "assistant" && isHeartbeatOkContent(stored.content)) {
-      sawHeartbeatAck = true;
-    }
-    if (sawHeartbeatMarker && sawHeartbeatAck) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-function filterSyntheticHeartbeatMessages(
-  messages: AgentMessage[],
-): { messages: AgentMessage[]; skipped: number } {
-  if (messages.length === 0) {
-    return { messages, skipped: 0 };
-  }
-
-  const skipIndexes = new Set<number>();
-  for (let index = 0; index < messages.length; index += 1) {
-    const stored = toStoredMessage(messages[index]!);
-    if (
-      stored.role === "user" &&
-      stored.content.trim().toLowerCase() === OPENCLAW_HEARTBEAT_POLL
-    ) {
-      skipIndexes.add(index);
-    }
-  }
-
-  for (let index = 0; index < messages.length; index += 1) {
-    const stored = toStoredMessage(messages[index]!);
-    if (stored.role !== "assistant" || !isHeartbeatOkContent(stored.content)) {
-      continue;
-    }
-
-    let turnStart = -1;
-    for (let cursor = index - 1; cursor >= 0; cursor -= 1) {
-      const previous = toStoredMessage(messages[cursor]!);
-      if (previous.role === "user") {
-        turnStart = cursor;
-        break;
-      }
-    }
-    if (turnStart < 0) {
-      continue;
-    }
-
-    const turnMessages = messages
-      .slice(turnStart, index + 1)
-      .map((message) => toStoredMessage(message));
-    if (!turnLooksLikeHeartbeatTurn(turnMessages)) {
-      continue;
-    }
-
-    for (let cursor = turnStart; cursor <= index; cursor += 1) {
-      skipIndexes.add(cursor);
-    }
-  }
-
-  if (skipIndexes.size === 0) {
-    return { messages, skipped: 0 };
-  }
-
-  return {
-    messages: messages.filter((_, index) => !skipIndexes.has(index)),
-    skipped: skipIndexes.size,
-  };
-}
-
-function turnLooksLikeHeartbeatTurn(turnMessages: Array<{ content: string }>): boolean {
-  return turnMessages.some((message) =>
-    message.content.toLowerCase().includes(HEARTBEAT_TURN_MARKER),
-  );
-}
 
 // ── Emergency fallback summarization ────────────────────────────────────────
